wip(tema): Inicia nova estrutura e adiciona componentes comuns
