+++ packages/agenciafmd/frontend/resources/views/components/privacy-policy-link.blade.php
@@ -1,0 +1,5 @@
+<x-frontend::link link="#"
+                  label="Termos e Condições"
+                  aria-label="Link Termos e Condições"
+                  title="Termos e Condições"
+                  class="text-reset text-decoration-none"/>
