+++ packages/agenciafmd/frontend/resources/views/html/tema.blade.php
@@ -1,2774 +1,2668 @@
-@extends('frontend::html.master')
+@extends('frontend::html.master', [
+'bodyClass' => '',
+'critical' => '',
+])
 
 @section('title', 'Tema')
 @section('description', 'Página para visualização de todos os componentes')
 
+@section('content')
+<main class="py-block">
+  <div class="container">
+    <div class="row gx-1 gy-block-small">
+      <div class="col-md-2">
+
+        <div class="navbar bd-example hstack gap-block-small position-sticky top-0">
+
+          <div>
+            <h2 class="fs-xlarge">Components</h2>
+            <ul id="nav-links-components"
+                class="navbar-nav flex-column">
+              @php
+              $componentsLinks = [
+              (object) ['name' => 'Buttons'],
+              (object) ['name' => 'Button Group'],
+              (object) ['name' => 'Button Share'],
+              (object) ['name' => 'Navbars'],
+              (object) ['name' => 'Navs - Tabs'],
+              ];
+              @endphp
+
+              @foreach($componentsLinks as $link)
+              <li class="nav-item">
+                <x-frontend::link link="#{{ Str::slug($link->name) }}"
+                                  label="{{ $link->name }}"
+                                  title="Ir para {{ $link->name }}"
+                                  class="px-0 nav-link js-scroll-top"/>
+              </li>
+              @endforeach
+            </ul>
+          </div>
+
+          <div>
+            <h2 class="fs-xlarge">Content</h2>
+            <ul id="nav-links-content"
+                class="navbar-nav flex-column">
+              @php
+              $contentLinks = [
+              (object) ['name' => 'Typography'],
+              (object) ['name' => 'Tables'],
+              ];
+              @endphp
+
+              @foreach($contentLinks as $link)
+              <li class="nav-item">
+                <x-frontend::link link="#{{ Str::slug($link->name) }}"
+                                  label="{{ $link->name }}"
+                                  title="Ir para {{ $link->name }}"
+                                  class="px-0 nav-link js-scroll-top"/>
+              </li>
+              @endforeach
+            </ul>
+          </div>
+
+          <div>
+            <h2 class="fs-xlarge">Forms</h2>
+            <ul id="nav-links-forms"
+                class="navbar-nav flex-column">
+              @php
+              $formsLinks = [
+              (object) ['name' => 'Form Control'],
+              (object) ['name' => 'Form Select'],
+              (object) ['name' => 'Checks / Radios'],
+              (object) ['name' => 'Range'],
+              (object) ['name' => 'Input Group'],
+              (object) ['name' => 'Floating Label'],
+              (object) ['name' => 'Validation'],
+              (object) ['name' => 'Livewire'],
+              ];
+              @endphp
+
+              @foreach($formsLinks as $link)
+              <li class="nav-item">
+                <x-frontend::link link="#{{ Str::slug($link->name) }}"
+                                  label="{{ $link->name }}"
+                                  title="Ir para {{ $link->name }}"
+                                  class="px-0 nav-link js-scroll-top"/>
+              </li>
+              @endforeach
+            </ul>
+          </div>
+
+        </div>
+
+      </div>
+      <div class="col-md-10">
+
+        <div class="vstack gap-block-small gap-md-block">
+          <div class="vstack gap-block-small gap-md-block"
+               data-bs-spy="scroll"
+               data-bs-target="#nav-links-components"
+               data-bs-smooth-scroll="true"
+               tabindex="0">
+
+            <div id="buttons"
+                 class="bd-example">
+              <h2 class="mb-block-small h6">Buttons</h2>
+              <div class="vstack gap-1">
+
+                <div>
+                  <button type="button"
+                          class="btn btn-primary">Primary
+                  </button>
+                  <button type="button"
+                          class="btn btn-secondary">Secondary
+                  </button>
+                  <button type="button"
+                          class="btn btn-success">Success
+                  </button>
+                  <button type="button"
+                          class="btn btn-info">Info
+                  </button>
+                  <button type="button"
+                          class="btn btn-warning">Warning
+                  </button>
+                  <button type="button"
+                          class="btn btn-danger">Danger
+                  </button>
+                  <button type="button"
+                          class="btn btn-light">Light
+                  </button>
+                  <button type="button"
+                          class="btn btn-dark">Dark
+                  </button>
+                  <button type="button"
+                          class="btn btn-link">Link
+                  </button>
+                </div>
+
+                <div>
+                  <button type="button"
+                          class="btn btn-primary disabled">Primary
+                  </button>
+                  <button type="button"
+                          class="btn btn-secondary disabled">Secondary
+                  </button>
+                  <button type="button"
+                          class="btn btn-success disabled">Success
+                  </button>
+                  <button type="button"
+                          class="btn btn-info disabled">Info
+                  </button>
+                  <button type="button"
+                          class="btn btn-warning disabled">Warning
+                  </button>
+                  <button type="button"
+                          class="btn btn-danger disabled">Danger
+                  </button>
+                  <button type="button"
+                          class="btn btn-light disabled">Light
+                  </button>
+                  <button type="button"
+                          class="btn btn-dark disabled">Dark
+                  </button>
+                  <button type="button"
+                          class="btn btn-link disabled">Link
+                  </button>
+                </div>
+
+                <div>
+                  <button type="button"
+                          class="btn btn-outline-primary">Primary
+                  </button>
+                  <button type="button"
+                          class="btn btn-outline-secondary">Secondary
+                  </button>
+                  <button type="button"
+                          class="btn btn-outline-success">Success
+                  </button>
+                  <button type="button"
+                          class="btn btn-outline-info">Info
+                  </button>
+                  <button type="button"
+                          class="btn btn-outline-warning">Warning
+                  </button>
+                  <button type="button"
+                          class="btn btn-outline-danger">Danger
+                  </button>
+                  <button type="button"
+                          class="btn btn-outline-light">Light
+                  </button>
+                  <button type="button"
+                          class="btn btn-outline-dark">Dark
+                  </button>
+                </div>
+
+                <div>
+                  <button type="button"
+                          class="btn btn-primary btn-lg">Large button
+                  </button>
+                  <button type="button"
+                          class="btn btn-primary">Default button
+                  </button>
+                  <button type="button"
+                          class="btn btn-primary btn-sm">Small button
+                  </button>
+                </div>
+              </div>
+            </div>
+
+            <div id="button-group"
+                 class="bd-example">
+              <h2 class="mb-block-small h6">Button Group</h2>
+              <div class="vstack gap-1">
+
+                <div>
+                  <div class="btn-group"
+                       role="group"
+                       aria-label="Basic checkbox toggle button group">
+                    <input type="checkbox"
+                           class="btn-check"
+                           id="btncheck1"
+                           checked
+                           autocomplete="off">
+                    <label class="btn btn-primary"
+                           for="btncheck1">Checkbox 1
+                    </label>
+                    <input type="checkbox"
+                           class="btn-check"
+                           id="btncheck2"
+                           autocomplete="off">
+                    <label class="btn btn-primary"
+                           for="btncheck2">Checkbox 2
+                    </label>
+                    <input type="checkbox"
+                           class="btn-check"
+                           id="btncheck3"
+                           autocomplete="off">
+                    <label class="btn btn-primary"
+                           for="btncheck3">Checkbox 3
+                    </label>
+                  </div>
+                </div>
+
+                <div>
+                  <div class="btn-group"
+                       role="group"
+                       aria-label="Basic radio toggle button group">
+                    <input type="radio"
+                           class="btn-check"
+                           name="btnradio"
+                           id="btnradio1"
+                           autocomplete="off"
+                           checked>
+                    <label class="btn btn-outline-primary"
+                           for="btnradio1">Radio 1
+                    </label>
+                    <input type="radio"
+                           class="btn-check"
+                           name="btnradio"
+                           id="btnradio2"
+                           autocomplete="off"
+                           checked>
+                    <label class="btn btn-outline-primary"
+                           for="btnradio2">Radio 2
+                    </label>
+                    <input type="radio"
+                           class="btn-check"
+                           name="btnradio"
+                           id="btnradio3"
+                           autocomplete="off"
+                           checked>
+                    <label class="btn btn-outline-primary"
+                           for="btnradio3">Radio 3
+                    </label>
+                  </div>
+                </div>
+
+                <div>
+                  <div class="btn-group-vertical">
+                    <button type="button"
+                            class="btn btn-primary">Button
+                    </button>
+                    <button type="button"
+                            class="btn btn-primary">Button
+                    </button>
+                    <button type="button"
+                            class="btn btn-primary">Button
+                    </button>
+                    <button type="button"
+                            class="btn btn-primary">Button
+                    </button>
+                    <button type="button"
+                            class="btn btn-primary">Button
+                    </button>
+                    <button type="button"
+                            class="btn btn-primary">Button
+                    </button>
+                  </div>
+                </div>
+
+                <div>
+                  <div class="btn-group"
+                       role="group"
+                       aria-label="Basic example">
+                    <button type="button"
+                            class="btn btn-secondary">Left
+                    </button>
+                    <button type="button"
+                            class="btn btn-secondary">Middle
+                    </button>
+                    <button type="button"
+                            class="btn btn-secondary">Right
+                    </button>
+                  </div>
+                </div>
+
+                <div>
+                  <div class="btn-toolbar"
+                       role="toolbar"
+                       aria-label="Toolbar with button groups">
+                    <div class="btn-group me-2"
+                         role="group"
+                         aria-label="First group">
+                      <button type="button"
+                              class="btn btn-secondary">
+                        1
+                      </button>
+                      <button type="button"
+                              class="btn btn-secondary">
+                        2
+                      </button>
+                      <button type="button"
+                              class="btn btn-secondary">
+                        3
+                      </button>
+                      <button type="button"
+                              class="btn btn-secondary">
+                        4
+                      </button>
+                    </div>
+                    <div class="btn-group me-2"
+                         role="group"
+                         aria-label="Second group">
+                      <button type="button"
+                              class="btn btn-secondary">
+                        5
+                      </button>
+                      <button type="button"
+                              class="btn btn-secondary">
+                        6
+                      </button>
+                      <button type="button"
+                              class="btn btn-secondary">
+                        7
+                      </button>
+                    </div>
+                    <div class="btn-group"
+                         role="group"
+                         aria-label="Third group">
+                      <button type="button"
+                              class="btn btn-secondary">
+                        8
+                      </button>
+                    </div>
+                  </div>
+                </div>
+
+              </div>
+            </div>
+
+            <div id="button-share"
+                 class="bd-example">
+              <h2 class="mb-block-small h6">Button Share</h2>
+              <div class="vstack gap-1">
+                <div>
+                  <x-frontend::link href="#"
+                                    role="button"
+                                    data-fmd-share-btn-fbidentification="123456789123456"
+                                    label="Compartilhar"
+                                    class="btn btn-secondary js-btn-share icon-link-hover-scale-center"
+                                    icon="ic-share"/>
+                </div>
+              </div>
+            </div>
+
+            <div id="navbars"
+                 class="bd-example">
+              <h2 class="mb-block-small h6">Navbars</h2>
+              <div class="vstack gap-1">
+                <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
+                  <div class="container-fluid">
+                    <a class="navbar-brand"
+                       href="#">Navbar
+                    </a>
+                    <button class="navbar-toggler"
+                            type="button"
+                            data-bs-toggle="collapse"
+                            data-bs-target="#navbarColor01"
+                            aria-controls="navbarColor01"
+                            aria-expanded="false"
+                            aria-label="Toggle navigation">
+                      <span class="navbar-toggler-icon"></span>
+                    </button>
+
+                    <div class="collapse navbar-collapse"
+                         id="navbarColor01">
+                      <ul class="navbar-nav me-auto">
+                        <li class="nav-item">
+                          <a class="nav-link active"
+                             href="#">Home
+                            <span class="visually-hidden">(current)</span>
+                          </a>
+                        </li>
+                        <li class="nav-item">
+                          <a class="nav-link"
+                             href="#">Features
+                          </a>
+                        </li>
+                        <li class="nav-item">
+                          <a class="nav-link"
+                             href="#">Pricing
+                          </a>
+                        </li>
+                        <li class="nav-item">
+                          <a class="nav-link"
+                             href="#">About
+                          </a>
+                        </li>
+                        <li class="nav-item dropdown">
+                          <a class="nav-link dropdown-toggle"
+                             data-bs-toggle="dropdown"
+                             href="#"
+                             role="button"
+                             aria-haspopup="true"
+                             aria-expanded="false">Dropdown
+                          </a>
+                          <div class="dropdown-menu">
+                            <a class="dropdown-item"
+                               href="#">Action
+                            </a>
+                            <a class="dropdown-item"
+                               href="#">Another action
+                            </a>
+                            <a class="dropdown-item"
+                               href="#">Something else here
+                            </a>
+                            <div class="dropdown-divider"></div>
+                            <a class="dropdown-item"
+                               href="#">Separated link
+                            </a>
+                          </div>
+                        </li>
+                      </ul>
+                      <form class="d-flex">
+                        <input class="form-control me-sm-2"
+                               type="text"
+                               placeholder="Search">
+                        <button class="btn btn-secondary my-2 my-sm-0"
+                                type="submit">Search
+                        </button>
+                      </form>
+                    </div>
+                  </div>
+                </nav>
+
+                <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
+                  <div class="container-fluid">
+                    <a class="navbar-brand"
+                       href="#">Navbar
+                    </a>
+                    <button class="navbar-toggler"
+                            type="button"
+                            data-bs-toggle="collapse"
+                            data-bs-target="#navbarColor02"
+                            aria-controls="navbarColor02"
+                            aria-expanded="false"
+                            aria-label="Toggle navigation">
+                      <span class="navbar-toggler-icon"></span>
+                    </button>
+
+                    <div class="collapse navbar-collapse"
+                         id="navbarColor02">
+                      <ul class="navbar-nav me-auto">
+                        <li class="nav-item">
+                          <a class="nav-link active"
+                             href="#">Home
+                            <span class="visually-hidden">(current)</span>
+                          </a>
+                        </li>
+                        <li class="nav-item">
+                          <a class="nav-link"
+                             href="#">Features
+                          </a>
+                        </li>
+                        <li class="nav-item">
+                          <a class="nav-link"
+                             href="#">Pricing
+                          </a>
+                        </li>
+                        <li class="nav-item">
+                          <a class="nav-link"
+                             href="#">About
+                          </a>
+                        </li>
+                        <li class="nav-item dropdown">
+                          <a class="nav-link dropdown-toggle"
+                             data-bs-toggle="dropdown"
+                             href="#"
+                             role="button"
+                             aria-haspopup="true"
+                             aria-expanded="false">Dropdown
+                          </a>
+                          <div class="dropdown-menu">
+                            <a class="dropdown-item"
+                               href="#">Action
+                            </a>
+                            <a class="dropdown-item"
+                               href="#">Another action
+                            </a>
+                            <a class="dropdown-item"
+                               href="#">Something else here
+                            </a>
+                            <div class="dropdown-divider"></div>
+                            <a class="dropdown-item"
+                               href="#">Separated link
+                            </a>
+                          </div>
+                        </li>
+                      </ul>
+                      <form class="d-flex">
+                        <input class="form-control me-sm-2"
+                               type="text"
+                               placeholder="Search">
+                        <button class="btn btn-secondary my-2 my-sm-0"
+                                type="submit">Search
+                        </button>
+                      </form>
+                    </div>
+                  </div>
+                </nav>
+
+                <nav class="navbar navbar-expand-lg navbar-light bg-light">
+                  <div class="container-fluid">
+                    <a class="navbar-brand"
+                       href="#">Navbar
+                    </a>
+                    <button class="navbar-toggler"
+                            type="button"
+                            data-bs-toggle="collapse"
+                            data-bs-target="#navbarColor03"
+                            aria-controls="navbarColor03"
+                            aria-expanded="false"
+                            aria-label="Toggle navigation">
+                      <span class="navbar-toggler-icon"></span>
+                    </button>
+
+                    <div class="collapse navbar-collapse"
+                         id="navbarColor03">
+                      <ul class="navbar-nav me-auto">
+                        <li class="nav-item">
+                          <a class="nav-link active"
+                             href="#">Home
+                            <span class="visually-hidden">(current)</span>
+                          </a>
+                        </li>
+                        <li class="nav-item">
+                          <a class="nav-link"
+                             href="#">Features
+                          </a>
+                        </li>
+                        <li class="nav-item">
+                          <a class="nav-link"
+                             href="#">Pricing
+                          </a>
+                        </li>
+                        <li class="nav-item">
+                          <a class="nav-link"
+                             href="#">About
+                          </a>
+                        </li>
+                        <li class="nav-item dropdown">
+                          <a class="nav-link dropdown-toggle"
+                             data-bs-toggle="dropdown"
+                             href="#"
+                             role="button"
+                             aria-haspopup="true"
+                             aria-expanded="false">Dropdown
+                          </a>
+                          <div class="dropdown-menu">
+                            <a class="dropdown-item"
+                               href="#">Action
+                            </a>
+                            <a class="dropdown-item"
+                               href="#">Another action
+                            </a>
+                            <a class="dropdown-item"
+                               href="#">Something else here
+                            </a>
+                            <div class="dropdown-divider"></div>
+                            <a class="dropdown-item"
+                               href="#">Separated link
+                            </a>
+                          </div>
+                        </li>
+                      </ul>
+                      <form class="d-flex">
+                        <input class="form-control me-sm-2"
+                               type="text"
+                               placeholder="Search">
+                        <button class="btn btn-secondary my-2 my-sm-0"
+                                type="submit">Search
+                        </button>
+                      </form>
+                    </div>
+                  </div>
+                </nav>
+
+                @include('frontend::html.partials.header')
+              </div>
+            </div>
+
+            <div id="navs-tabs"
+                 class="bd-example">
+              <h2 class="mb-block-small h6">Navs</h2>
+              <div class="vstack gap-1">
+
+                <div>
+                  <ul class="nav">
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Active"
+                                        class="nav-link active"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link disabled"/>
+                    </li>
+                  </ul>
+                </div>
+
+                <div>
+                  <ul class="nav justify-content-center">
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Active"
+                                        class="nav-link active"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link disabled"/>
+                    </li>
+                  </ul>
+                </div>
+
+                <div>
+                  <ul class="nav justify-content-end">
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Active"
+                                        class="nav-link active"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link disabled"/>
+                    </li>
+                  </ul>
+                </div>
+
+                <div>
+                  <ul class="nav vstack align-content-start">
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Active"
+                                        class="nav-link active"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link disabled"/>
+                    </li>
+                  </ul>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example nav-tabs</h2>
+
+                  <ul class="nav nav-tabs">
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Active"
+                                        class="nav-link active"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link disabled"/>
+                    </li>
+                  </ul>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example nav-tabs with dropdown</h2>
+
+                  <ul class="nav nav-tabs">
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Active"
+                                        class="nav-link active"/>
+                    </li>
+                    <li class="nav-item dropdown">
+                      <x-frontend::link link="#"
+                                        title="Ir para Contato"
+                                        aria-label="Link para Contato"
+                                        label="Produtos"
+                                        role="button"
+                                        data-bs-toggle="dropdown"
+                                        aria-expanded="false"
+                                        class="nav-link dropdown-toggle"/>
+                      <ul class="dropdown-menu">
+                        <li>
+                          <x-frontend::link link="/html/"
+                                            title="Ir para Produto 1"
+                                            aria-label="Link para Produto 1"
+                                            label="Produto 1"
+                                            class="dropdown-item"/>
+                        </li>
+                        <li>
+                          <x-frontend::link link="/html/"
+                                            title="Ir para Produto 2"
+                                            aria-label="Link para Produto 2"
+                                            label="Produto 2"
+                                            class="dropdown-item"/>
+                        </li>
+                        <li>
+                          <x-frontend::link link="/html/"
+                                            title="Ir para Produto 3"
+                                            aria-label="Link para Produto 3"
+                                            label="Produto 3"
+                                            class="dropdown-item"/>
+                        </li>
+                      </ul>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link disabled"/>
+                    </li>
+                  </ul>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example nav-pills</h2>
+
+                  <ul class="nav nav-pills">
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Active"
+                                        class="nav-link active"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link disabled"/>
+                    </li>
+                  </ul>
+
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example nav-pills with dropdown</h2>
+
+                  <ul class="nav nav-pills">
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Active"
+                                        class="nav-link active"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item dropdown">
+                      <x-frontend::link link="#"
+                                        title="Ir para Contato"
+                                        aria-label="Link para Contato"
+                                        label="Produtos"
+                                        role="button"
+                                        data-bs-toggle="dropdown"
+                                        aria-expanded="false"
+                                        class="nav-link dropdown-toggle"/>
+                      <ul class="dropdown-menu">
+                        <li>
+                          <x-frontend::link link="/html/"
+                                            title="Ir para Produto 1"
+                                            aria-label="Link para Produto 1"
+                                            label="Produto 1"
+                                            class="dropdown-item"/>
+                        </li>
+                        <li>
+                          <x-frontend::link link="/html/"
+                                            title="Ir para Produto 2"
+                                            aria-label="Link para Produto 2"
+                                            label="Produto 2"
+                                            class="dropdown-item"/>
+                        </li>
+                        <li>
+                          <x-frontend::link link="/html/"
+                                            title="Ir para Produto 3"
+                                            aria-label="Link para Produto 3"
+                                            label="Produto 3"
+                                            class="dropdown-item"/>
+                        </li>
+                      </ul>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link disabled"/>
+                    </li>
+                  </ul>
+
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example nav-underline</h2>
+
+                  <ul class="nav nav-underline">
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Active"
+                                        class="nav-link active"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link disabled"/>
+                    </li>
+                  </ul>
+
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example nav-pills fill and justify</h2>
+
+                  <ul class="nav nav-pills nav-fill">
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Active"
+                                        class="nav-link active"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link disabled"/>
+                    </li>
+                  </ul>
+
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example nav-pills justify</h2>
+
+                  <ul class="nav nav-pills nav-justified">
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Active"
+                                        class="nav-link active"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link"/>
+                    </li>
+                    <li class="nav-item">
+                      <x-frontend::link link="#"
+                                        aria-current="page"
+                                        label="Link"
+                                        class="nav-link disabled"/>
+                    </li>
+                  </ul>
+
+                </div>
+
+              </div>
+            </div>
+
+          </div>
+
+          <div class="vstack gap-block-small gap-md-block"
+               data-bs-spy="scroll"
+               data-bs-target="#nav-links-content"
+               data-bs-smooth-scroll="true"
+               tabindex="0">
+
+            <div id="typography"
+                 class="bd-example">
+              <h2 class="mb-block-small h6">Typography</h2>
+              <div class="vstack gap-1">
+
+                <div>
+                  <h1>Heading 1</h1>
+                  <h2>Heading 2</h2>
+                  <h3>Heading 3</h3>
+                  <h4>Heading 4</h4>
+                  <h5>Heading 5</h5>
+                  <h6>Heading 6</h6>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example text-muted and lead</h2>
+                  <h3>
+                    Heading
+                    <small class="text-muted">with muted text</small>
+                  </h3>
+                  <p class="lead">
+                    Vivamus sagittis lacus vel augue laoreet rutrum faucibus
+                    dolor auctor.
+                  </p>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example body text</h2>
+                  <p>Nullam quis risus eget
+                    <x-frontend::link link="#"
+                                      label="urna mollis ornare"/>
+                     vel eu leo. Cum sociis natoque
+                     penatibus et magnis dis parturient montes, nascetur
+                     ridiculus mus. Nullam id dolor id nibh
+                     ultricies vehicula.
+                  </p>
+                  <p>
+                    <small>This line of text is meant to be treated as fine
+                           print.
+                    </small>
+                  </p>
+                  <p>The following is <strong>rendered as bold text</strong>.
+                  </p>
+                  <p>The following is <em>rendered as italicized text</em>.
+                  </p>
+                  <p>An abbreviation of the word attribute is
+                    <abbr title="attribute">attr</abbr>.
+                  </p>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example list-unstyled</h2>
+                  <ul class="list-unstyled">
+                    <li>This is a list.</li>
+                    <li>It appears completely unstyled.</li>
+                    <li>Structurally, it's still a list.</li>
+                    <li>However, this style only applies to immediate child
+                        elements.
+                    </li>
+                    <li>Nested lists:
+                      <ul>
+                        <li>are unaffected by this style</li>
+                        <li>will still show a bullet</li>
+                        <li>and have appropriate left margin</li>
+                      </ul>
+                    </li>
+                    <li>This may still come in handy in some situations.</li>
+                  </ul>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example list-inline</h2>
+                  <ul class="list-inline">
+                    <li class="list-inline-item">This is a list item.</li>
+                    <li class="list-inline-item">And another one.</li>
+                    <li class="list-inline-item">But they're displayed inline.
+                    </li>
+                  </ul>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example blockquotes</h2>
+                  <div class="row gy-block-small">
+                    <div class="col-md-4">
+                      <figure>
+                        <blockquote class="blockquote">
+                          <p class="mb-0">
+                            Lorem ipsum dolor sit amet, consectetur adipiscing
+                            elit. Integer posuere
+                            erat a ante.
+                          </p>
+
+                        </blockquote>
+                        <figcaption class="blockquote-footer">
+                          Someone famous in <cite title="Source Title">
+                          Source Title</cite>
+                        </figcaption>
+                      </figure>
+                    </div>
+                    <div class="col-md-4">
+                      <figure class="text-center">
+                        <blockquote class="blockquote">
+                          <p class="mb-0">
+                            Lorem ipsum dolor sit amet, consectetur adipiscing
+                            elit. Integer posuere
+                            erat a ante.
+                          </p>
+
+                        </blockquote>
+                        <figcaption class="blockquote-footer">
+                          Someone famous in <cite title="Source Title">
+                          Source Title</cite>
+                        </figcaption>
+                      </figure>
+                    </div>
+                    <div class="col-md-4">
+                      <figure class="text-end">
+                        <blockquote class="blockquote">
+                          <p class="mb-0">
+                            Lorem ipsum dolor sit amet, consectetur adipiscing
+                            elit. Integer posuere
+                            erat a ante.
+                          </p>
+
+                        </blockquote>
+                        <figcaption class="blockquote-footer">
+                          Someone famous in <cite title="Source Title">
+                          Source Title</cite>
+                        </figcaption>
+                      </figure>
+                    </div>
+                  </div>
+                </div>
+
+              </div>
+            </div>
+
+            <div id="tables"
+                 class="bd-example">
+              <h2 class="mb-block-small h6">Tables</h2>
+              <div class="vstack gap-1">
+                <table class="table table-hover">
+                  <thead>
+                    <tr>
+                      <th scope="col">Type</th>
+                      <th scope="col">Column heading</th>
+                      <th scope="col">Column heading</th>
+                      <th scope="col">Column heading</th>
+                    </tr>
+                  </thead>
+                  <tbody>
+                    <tr class="table-active">
+                      <th scope="row">Active</th>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                    </tr>
+                    <tr>
+                      <th scope="row">Default</th>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                    </tr>
+                    <tr class="table-primary">
+                      <th scope="row">Primary</th>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                    </tr>
+                    <tr class="table-secondary">
+                      <th scope="row">Secondary</th>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                    </tr>
+                    <tr class="table-success">
+                      <th scope="row">Success</th>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                    </tr>
+                    <tr class="table-danger">
+                      <th scope="row">Danger</th>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                    </tr>
+                    <tr class="table-warning">
+                      <th scope="row">Warning</th>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                    </tr>
+                    <tr class="table-info">
+                      <th scope="row">Info</th>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                    </tr>
+                    <tr class="table-light">
+                      <th scope="row">Light</th>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                    </tr>
+                    <tr class="table-dark">
+                      <th scope="row">Dark</th>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                      <td>Column content</td>
+                    </tr>
+                  </tbody>
+                </table>
+              </div>
+            </div>
+
+          </div>
+
+          <div class="vstack gap-block-small gap-md-block"
+               data-bs-spy="scroll"
+               data-bs-target="#nav-links-forms"
+               data-bs-smooth-scroll="true"
+               tabindex="0">
+
+            <div id="form-control"
+                 class="bd-example">
+
+              <h2 class="mb-block-small h6">Form Control</h2>
+              <div class="vstack gap-1">
+
+                <div>
+                  <div class="row gy-block-small">
+                    <div class="col-md-6">
+                      <label for="exampleFormControlInput1"
+                             class="form-label">Example input
+                      </label>
+                      <input type="text"
+                             class="form-control"
+                             id="exampleFormControlInput1"
+                             placeholder="">
+                    </div>
+                    <div class="col-md-6">
+                      <label for="exampleFormControlTextarea1"
+                             class="form-label">Example textarea
+                      </label>
+                      <textarea class="form-control"
+                                id="exampleFormControlTextarea1"
+                                rows="3"></textarea>
+                    </div>
+                  </div>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Sizing</h2>
+
+                  <div class="row gy-block-small">
+                    <div class="col-md-4">
+                      <label class="col-form-label col-form-label-sm"
+                             for="inputSmall">Small input
+                      </label>
+                      <input class="form-control form-control-sm"
+                             type="text"
+                             placeholder=".form-control-sm"
+                             id="inputSmall">
+                    </div>
+                    <div class="col-md-4">
+                      <label class="col-form-label"
+                             for="inputDefault">Default input
+                      </label>
+                      <input type="text"
+                             class="form-control"
+                             placeholder="Default input"
+                             id="inputDefault">
+                    </div>
+                    <div class="col-md-4">
+                      <label class="col-form-label col-form-label-lg"
+                             for="inputLarge">Large input
+                      </label>
+                      <input class="form-control form-control-lg"
+                             type="text"
+                             placeholder=".form-control-lg"
+                             id="inputLarge">
+                    </div>
+                  </div>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Form text</h2>
+
+                  <div class="row align-items-center">
+                    <div class="col-12">
+                      <label for="inputPassword6"
+                             class="col-form-label">Password
+                      </label>
+                    </div>
+                    <div class="col-12">
+                      <input type="password"
+                             id="inputPassword6"
+                             class="form-control"
+                             aria-describedby="passwordHelpInline">
+                    </div>
+                    <div class="col-12">
+                      <span id="passwordHelpInline"
+                            class="form-text">
+                        Must be 8-20 characters long.
+                      </span>
+                    </div>
+                  </div>
+
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Disabled Input</h2>
+
+                  <div>
+                    <label class="form-label"
+                           for="disabledInput">Disabled input
+                    </label>
+                    <input class="form-control"
+                           id="disabledInput"
+                           type="text"
+                           placeholder="Disabled input here..."
+                           disabled>
+                  </div>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Readonly Input</h2>
+
+                  <div>
+                    <label class="form-label"
+                           for="readOnlyInput">Readonly input
+                    </label>
+                    <input class="form-control"
+                           id="readOnlyInput"
+                           type="text"
+                           placeholder="Readonly input here..."
+                           readonly>
+                  </div>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example File Input</h2>
+
+                  <div class="row gy-1">
+                    <div class="col-md-6">
+                      <label for="formFileSm"
+                             class="form-label">Small file input example
+                      </label>
+                      <input class="form-control form-control-sm"
+                             id="formFileSm"
+                             type="file">
+                    </div>
+                    <div class="col-md-6">
+                      <label for="formFile"
+                             class="form-label">Default file input example
+                      </label>
+                      <input class="form-control"
+                             type="file"
+                             id="formFile">
+                    </div>
+                    <div class="col-md-6">
+                      <label for="formFileLg"
+                             class="form-label">Large file input example
+                      </label>
+                      <input class="form-control form-control-lg"
+                             id="formFileLg"
+                             type="file">
+                    </div>
+                    <div class="col-md-6">
+                      <label for="formFileMultiple"
+                             class="form-label">Multiple files input example
+                      </label>
+                      <input class="form-control"
+                             type="file"
+                             id="formFileMultiple"
+                             multiple>
+                    </div>
+                    <div class="col-md-6">
+                      <label for="formFileDisabled"
+                             class="form-label">Disabled file input example
+                      </label>
+                      <input class="form-control"
+                             type="file"
+                             id="formFileDisabled"
+                             disabled>
+                    </div>
+                  </div>
+                </div>
+
+              </div>
+
+            </div>
+
+            <div id="form-select"
+                 class="bd-example">
+
+              <h2 class="h6">Form Select</h2>
+              <div class="vstack gap-1">
+
+                <div>
+                  <div class="row gy-block-small">
+                    <div class="col-md-6">
+                      <label for="defaultSelect"
+                             class="form-label mt-1h">Example select
+                      </label>
+                      <select class="form-select"
+                              id="defaultSelect"
+                              aria-label="Default select example">
+                        <option selected>Open this select menu</option>
+                        <option value="1">One</option>
+                        <option value="2">Two</option>
+                        <option value="3">Three</option>
+                      </select>
+                    </div>
+                    <div class="col-md-6">
+                      <label for="multipleSelect"
+                             class="form-label mt-1h">Example multiple select
+                      </label>
+                      <select class="form-select"
+                              id="multipleSelect"
+                              multiple
+                              aria-label="Multiple select example">
+                        <option selected>Open this select menu</option>
+                        <option value="1">One</option>
+                        <option value="2">Two</option>
+                        <option value="3">Three</option>
+                      </select>
+                    </div>
+                  </div>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Sizing</h2>
+
+                  <div class="row gy-block-small">
+                    <div class="col-md-4">
+                      <label class="col-form-label col-form-label-sm"
+                             for="selectDefault">Small select
+                      </label>
+                      <select class="form-select form-select-sm"
+                              aria-label="Small select example">
+                        <option selected>Open this select menu</option>
+                        <option value="1">One</option>
+                        <option value="2">Two</option>
+                        <option value="3">Three</option>
+                      </select>
+                    </div>
+                    <div class="col-md-4">
+                      <label class="col-form-label"
+                             for="selectDefault">Default select
+                      </label>
+                      <select class="form-select"
+                              id="selectDefault"
+                              aria-label="Large select example">
+                        <option selected>Open this select menu</option>
+                        <option value="1">One</option>
+                        <option value="2">Two</option>
+                        <option value="3">Three</option>
+                      </select>
+                    </div>
+                    <div class="col-md-4">
+                      <label class="col-form-label col-form-label-lg"
+                             for="selectLarge">Large select
+                      </label>
+                      <select class="form-select form-select-lg"
+                              id="selectLarge"
+                              aria-label="Large select example">
+                        <option selected>Open this select menu</option>
+                        <option value="1">One</option>
+                        <option value="2">Two</option>
+                        <option value="3">Three</option>
+                      </select>
+                    </div>
+                  </div>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Disabled Input</h2>
+
+                  <div>
+                    <label for="disabledSelect"
+                           class="form-label mt-1h">Disabled select
+                    </label>
+                    <select class="form-select"
+                            id="disabledSelect"
+                            aria-label="Disabled select example"
+                            disabled>
+                      <option selected>Open this select menu</option>
+                      <option value="1">One</option>
+                      <option value="2">Two</option>
+                      <option value="3">Three</option>
+                    </select>
+                  </div>
+                </div>
+
+              </div>
+
+            </div>
+
+            <div id="checks-radios"
+                 class="bd-example">
+
+              <h2 class="h6">Checkbox</h2>
+              <div class="vstack gap-1">
+
+                <div>
+                  <div class="form-check">
+                    <input class="form-check-input"
+                           type="checkbox"
+                           value=""
+                           id="flexCheckDefault">
+                    <label class="form-check-label"
+                           for="flexCheckDefault">
+                      Default checkbox
+                    </label>
+                  </div>
+                  <div class="form-check">
+                    <input class="form-check-input"
+                           type="checkbox"
+                           value=""
+                           id="flexCheckChecked"
+                           checked>
+                    <label class="form-check-label"
+                           for="flexCheckChecked">
+                      Checked checkbox
+                    </label>
+                  </div>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Disabled Checkbox</h2>
+
+                  <div class="form-check">
+                    <input class="form-check-input"
+                           type="checkbox"
+                           value=""
+                           id="flexCheckDisabled"
+                           disabled>
+                    <label class="form-check-label"
+                           for="flexCheckDisabled">
+                      Disabled checkbox
+                    </label>
+                  </div>
+                  <div class="form-check">
+                    <input class="form-check-input"
+                           type="checkbox"
+                           value=""
+                           id="flexCheckCheckedDisabled"
+                           checked
+                           disabled>
+                    <label class="form-check-label"
+                           for="flexCheckCheckedDisabled">
+                      Disabled checked checkbox
+                    </label>
+                  </div>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Radio</h2>
+
+                  <div class="form-check">
+                    <input class="form-check-input"
+                           type="radio"
+                           name="flexRadioDefault"
+                           id="flexRadioDefault1">
+                    <label class="form-check-label"
+                           for="flexRadioDefault1">
+                      Default radio
+                    </label>
+                  </div>
+                  <div class="form-check">
+                    <input class="form-check-input"
+                           type="radio"
+                           name="flexRadioDefault"
+                           id="flexRadioDefault2"
+                           checked>
+                    <label class="form-check-label"
+                           for="flexRadioDefault2">
+                      Default checked radio
+                    </label>
+                  </div>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Disabled Radio</h2>
+
+                  <div class="form-check">
+                    <input class="form-check-input"
+                           type="radio"
+                           name="flexRadioDisabled"
+                           id="flexRadioDisabled"
+                           disabled>
+                    <label class="form-check-label"
+                           for="flexRadioDisabled">
+                      Disabled radio
+                    </label>
+                  </div>
+                  <div class="form-check">
+                    <input class="form-check-input"
+                           type="radio"
+                           name="flexRadioDisabled"
+                           id="flexRadioCheckedDisabled"
+                           checked
+                           disabled>
+                    <label class="form-check-label"
+                           for="flexRadioCheckedDisabled">
+                      Disabled checked radio
+                    </label>
+                  </div>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example switches</h2>
+
+                  <div class="form-check form-switch">
+                    <input class="form-check-input"
+                           type="checkbox"
+                           role="switch"
+                           id="flexSwitchCheckDefault">
+                    <label class="form-check-label"
+                           for="flexSwitchCheckDefault">
+                      Default switch checkbox input
+                    </label>
+                  </div>
+                  <div class="form-check form-switch">
+                    <input class="form-check-input"
+                           type="checkbox"
+                           role="switch"
+                           id="flexSwitchCheckChecked"
+                           checked>
+                    <label class="form-check-label"
+                           for="flexSwitchCheckChecked">
+                      Checked switch checkbox input
+                    </label>
+                  </div>
+                  <div class="form-check form-switch">
+                    <input class="form-check-input"
+                           type="checkbox"
+                           role="switch"
+                           id="flexSwitchCheckDisabled"
+                           disabled>
+                    <label class="form-check-label"
+                           for="flexSwitchCheckDisabled">
+                      Disabled switch checkbox input
+                    </label>
+                  </div>
+                  <div class="form-check form-switch">
+                    <input class="form-check-input"
+                           type="checkbox"
+                           role="switch"
+                           id="flexSwitchCheckCheckedDisabled"
+                           checked
+                           disabled>
+                    <label class="form-check-label"
+                           for="flexSwitchCheckCheckedDisabled">
+                      Disabled checked switch checkbox input
+                    </label>
+                  </div>
+
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example checkbox inline</h2>
+
+                  <div class="form-check form-check-inline">
+                    <input class="form-check-input"
+                           type="checkbox"
+                           id="inlineCheckbox1"
+                           value="option1">
+                    <label class="form-check-label"
+                           for="inlineCheckbox1">1
+                    </label>
+                  </div>
+                  <div class="form-check form-check-inline">
+                    <input class="form-check-input"
+                           type="checkbox"
+                           id="inlineCheckbox2"
+                           value="option2">
+                    <label class="form-check-label"
+                           for="inlineCheckbox2">2
+                    </label>
+                  </div>
+                  <div class="form-check form-check-inline">
+                    <input class="form-check-input"
+                           type="checkbox"
+                           id="inlineCheckbox3"
+                           value="option3"
+                           disabled>
+                    <label class="form-check-label"
+                           for="inlineCheckbox3">3 (disabled)
+                    </label>
+                  </div>
+
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example radio inline</h2>
+
+                  <div class="form-check form-check-inline">
+                    <input class="form-check-input"
+                           type="radio"
+                           name="inlineRadioOptions"
+                           id="inlineRadio1"
+                           value="option1">
+                    <label class="form-check-label"
+                           for="inlineRadio1">1
+                    </label>
+                  </div>
+                  <div class="form-check form-check-inline">
+                    <input class="form-check-input"
+                           type="radio"
+                           name="inlineRadioOptions"
+                           id="inlineRadio2"
+                           value="option2">
+                    <label class="form-check-label"
+                           for="inlineRadio2">2
+                    </label>
+                  </div>
+                  <div class="form-check form-check-inline">
+                    <input class="form-check-input"
+                           type="radio"
+                           name="inlineRadioOptions"
+                           id="inlineRadio3"
+                           value="option3"
+                           disabled>
+                    <label class="form-check-label"
+                           for="inlineRadio3">3 (disabled)
+                    </label>
+                  </div>
+
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example checkbox toggle buttons</h2>
+
+                  <input type="checkbox"
+                         class="btn-check"
+                         id="btn-check"
+                         autocomplete="off">
+                  <label class="btn btn-primary"
+                         for="btn-check">Single toggle
+                  </label>
+
+                  <input type="checkbox"
+                         class="btn-check"
+                         id="btn-check-2"
+                         checked
+                         autocomplete="off">
+                  <label class="btn btn-primary"
+                         for="btn-check-2">Checked
+                  </label>
+
+                  <input type="checkbox"
+                         class="btn-check"
+                         id="btn-check-3"
+                         autocomplete="off"
+                         disabled>
+                  <label class="btn btn-primary"
+                         for="btn-check-3">Disabled
+                  </label>
+
+                  <input type="checkbox"
+                         class="btn-check"
+                         id="btn-check-4"
+                         autocomplete="off">
+                  <label class="btn"
+                         for="btn-check-4">Single toggle
+                  </label>
+
+                  <input type="checkbox"
+                         class="btn-check"
+                         id="btn-check-5"
+                         checked
+                         autocomplete="off">
+                  <label class="btn"
+                         for="btn-check-5">Checked
+                  </label>
+
+                  <input type="checkbox"
+                         class="btn-check"
+                         id="btn-check-6"
+                         autocomplete="off"
+                         disabled>
+                  <label class="btn"
+                         for="btn-check-6">Disabled
+                  </label>
+
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example radio toggle buttons</h2>
+
+                  <input type="radio"
+                         class="btn-check"
+                         name="options"
+                         id="option1"
+                         autocomplete="off"
+                         checked>
+                  <label class="btn btn-secondary"
+                         for="option1">Checked
+                  </label>
+
+                  <input type="radio"
+                         class="btn-check"
+                         name="options"
+                         id="option2"
+                         autocomplete="off">
+                  <label class="btn btn-secondary"
+                         for="option2">Radio
+                  </label>
+
+                  <input type="radio"
+                         class="btn-check"
+                         name="options"
+                         id="option3"
+                         autocomplete="off"
+                         disabled>
+                  <label class="btn btn-secondary"
+                         for="option3">Disabled
+                  </label>
+
+                  <input type="radio"
+                         class="btn-check"
+                         name="options"
+                         id="option4"
+                         autocomplete="off">
+                  <label class="btn btn-secondary"
+                         for="option4">Radio
+                  </label>
+
+                  <input type="radio"
+                         class="btn-check"
+                         name="options-base"
+                         id="option5"
+                         autocomplete="off"
+                         checked>
+                  <label class="btn"
+                         for="option5">Checked
+                  </label>
+
+                  <input type="radio"
+                         class="btn-check"
+                         name="options-base"
+                         id="option6"
+                         autocomplete="off">
+                  <label class="btn"
+                         for="option6">Radio
+                  </label>
+
+                  <input type="radio"
+                         class="btn-check"
+                         name="options-base"
+                         id="option7"
+                         autocomplete="off"
+                         disabled>
+                  <label class="btn"
+                         for="option7">Disabled
+                  </label>
+
+                  <input type="radio"
+                         class="btn-check"
+                         name="options-base"
+                         id="option8"
+                         autocomplete="off">
+                  <label class="btn"
+                         for="option8">Radio
+                  </label>
+
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example outlined styles</h2>
+
+                  <input type="checkbox"
+                         class="btn-check"
+                         id="btn-check-outlined"
+                         autocomplete="off">
+                  <label class="btn btn-outline-primary"
+                         for="btn-check-outlined">Single toggle
+                  </label>
+
+                  <input type="checkbox"
+                         class="btn-check"
+                         id="btn-check-2-outlined"
+                         checked
+                         autocomplete="off">
+                  <label class="btn btn-outline-secondary"
+                         for="btn-check-2-outlined">Checked
+                  </label>
+
+                  <input type="radio"
+                         class="btn-check"
+                         name="options-outlined"
+                         id="success-outlined"
+                         autocomplete="off"
+                         checked>
+                  <label class="btn btn-outline-success"
+                         for="success-outlined">Checked success radio
+                  </label>
+
+                  <input type="radio"
+                         class="btn-check"
+                         name="options-outlined"
+                         id="danger-outlined"
+                         autocomplete="off">
+                  <label class="btn btn-outline-danger"
+                         for="danger-outlined">Danger radio
+                  </label>
+
+                </div>
+
+              </div>
+
+            </div>
+
+            <div id="range"
+                 class="bd-example">
+
+              <h2 class="h6">Range</h2>
+              <div class="vstack gap-1">
+
+                <div>
+                  <label for="customRange1"
+                         class="form-label">Example range
+                  </label>
+                  <input type="range"
+                         class="form-range"
+                         id="customRange1">
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Disabled Range</h2>
+
+                  <label for="disabledRange"
+                         class="form-label">Disabled range
+                  </label>
+                  <input type="range"
+                         class="form-range"
+                         id="disabledRange"
+                         disabled>
+
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Min and max Range</h2>
+
+                  <label for="customRange2"
+                         class="form-label">Example range
+                  </label>
+                  <input type="range"
+                         class="form-range"
+                         min="0"
+                         max="5"
+                         id="customRange2">
+
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Steps Range</h2>
+
+                  <label for="customRange3"
+                         class="form-label">Example range
+                  </label>
+                  <input type="range"
+                         class="form-range"
+                         min="0"
+                         max="5"
+                         step="0.5"
+                         id="customRange3">
+
+                </div>
+
+              </div>
+            </div>
+
+            <div id="input-group"
+                 class="bd-example">
+
+              <h2 class="h6">Input Group</h2>
+              <div class="vstack gap-1">
+
+                <div class="input-group">
+                  <span class="input-group-text"
+                        id="basic-addon1">@
+                  </span>
+                  <input type="text"
+                         class="form-control"
+                         placeholder="Username"
+                         aria-label="Username"
+                         aria-describedby="basic-addon1">
+                </div>
+
+                <div class="input-group">
+                  <input type="text"
+                         class="form-control"
+                         placeholder="Recipient's username"
+                         aria-label="Recipient's username"
+                         aria-describedby="basic-addon2">
+                  <span class="input-group-text"
+                        id="basic-addon2">@example.com
+                  </span>
+                </div>
+
+                <div class="">
+                  <label for="basic-url"
+                         class="form-label">
+                    Your vanity URL
+                  </label>
+                  <div class="input-group">
+                    <span class="input-group-text"
+                          id="basic-addon3">https://example.com/users/
+                    </span>
+                    <input type="text"
+                           class="form-control"
+                           id="basic-url"
+                           aria-describedby="basic-addon3 basic-addon4">
+                  </div>
+                  <div class="form-text"
+                       id="basic-addon4">
+                    Example help text goes outside the input group.
+                  </div>
+                </div>
+
+                <div class="input-group">
+                  <span class="input-group-text">$</span>
+                  <input type="text"
+                         class="form-control"
+                         aria-label="Amount (to the nearest dollar)">
+                  <span class="input-group-text">.00</span>
+                </div>
+
+                <div class="input-group">
+                  <input type="text"
+                         class="form-control"
+                         placeholder="Username"
+                         aria-label="Username">
+                  <span class="input-group-text">@</span>
+                  <input type="text"
+                         class="form-control"
+                         placeholder="Server"
+                         aria-label="Server">
+                </div>
+
+                <div class="input-group">
+                  <span class="input-group-text">With textarea</span>
+                  <textarea class="form-control"
+                            aria-label="With textarea"></textarea>
+                </div>
+
+              </div>
+            </div>
+
+            <div id="floating-label"
+                 class="bd-example">
+
+              <h2 class="h6">Floating label</h2>
+              <div class="vstack gap-1">
+
+                <div>
+                  <div class="row g-1">
+                    <div class="col-md-6">
+                      <div class="form-floating">
+                        <input type="email"
+                               class="form-control"
+                               id="floatingInput"
+                               placeholder="name@example.com">
+                        <label for="floatingInput">Email address</label>
+                      </div>
+                    </div>
+                    <div class="col-md-6">
+                      <div class="form-floating">
+                        <input type="password"
+                               class="form-control"
+                               id="floatingPassword"
+                               placeholder="Password">
+                        <label for="floatingPassword">Password</label>
+                      </div>
+                    </div>
+                  </div>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Textarea</h2>
+                  <div class="form-floating">
+                    <textarea class="form-control"
+                              placeholder="Leave a comment here"
+                              id="floatingTextarea2"></textarea>
+                    <label for="floatingTextarea2">Comments</label>
+                  </div>
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Select</h2>
+
+                  <div class="form-floating">
+                    <select class="form-select"
+                            id="floatingSelect"
+                            aria-label="Floating label select example">
+                      <option selected></option>
+                      <option value="1">One</option>
+                      <option value="2">Two</option>
+                      <option value="3">Three</option>
+                    </select>
+                    <label for="floatingSelect">Works with selects</label>
+                  </div>
+
+                </div>
+
+                <div>
+                  <hr>
+                  <h2 class="fs-xlarge">Example Disabled</h2>
+
+                  <div class="row g-1">
+                    <div class="col-md-6">
+                      <div class="form-floating">
+                        <input type="email"
+                               class="form-control"
+                               id="floatingInputDisabled"
+                               placeholder="name@example.com"
+                               disabled>
+                        <label for="floatingInputDisabled">Email address</label>
+                      </div>
+                    </div>
+                    <div class="col-md-6">
+                      <div class="form-floating">
+                        <textarea class="form-control"
+                                  placeholder="Leave a comment here"
+                                  id="floatingTextareaDisabled"
+                                  disabled></textarea>
+                        <label for="floatingTextareaDisabled">Comments</label>
+                      </div>
+                    </div>
+                    <div class="col-md-6">
+                      <div class="form-floating">
+                        <textarea class="form-control"
+                                  placeholder="Leave a comment here"
+                                  id="floatingTextarea2Disabled"
+                                  style="height: 100px"
+                                  disabled>Disabled textarea with some text inside</textarea>
+                        <label for="floatingTextarea2Disabled">Comments</label>
+                      </div>
+                    </div>
+                    <div class="col-md-6">
+                      <div class="form-floating">
+                        <select class="form-select"
+                                id="floatingSelectDisabled"
+                                aria-label="Floating label disabled select example"
+                                disabled>
+                          <option selected>Open this select menu</option>
+                          <option value="1">One</option>
+                          <option value="2">Two</option>
+                          <option value="3">Three</option>
+                        </select>
+                        <label for="floatingSelectDisabled">Works with selects
+                        </label>
+                      </div>
+                    </div>
+                  </div>
+                </div>
+
+              </div>
+
+            </div>
+
+            <div id="validation"
+                 class="bd-example">
+
+              <h2 class="h6">Validation</h2>
+              <div class="vstack gap-1">
+
+                <form class="needs-validation"
+                      novalidate>
+                  <div class="row g-1">
+                    <div class="col-md-4">
+                      <label for="validationCustom01"
+                             class="form-label">First name
+                      </label>
+                      <input type="text"
+                             class="form-control"
+                             id="validationCustom01"
+                             value="Mark"
+                             required>
+                      <div class="valid-feedback">
+                        Looks good!
+                      </div>
+                    </div>
+                    <div class="col-md-4">
+                      <label for="validationCustom02"
+                             class="form-label">Last name
+                      </label>
+                      <input type="text"
+                             class="form-control"
+                             id="validationCustom02"
+                             value="Otto"
+                             required>
+                      <div class="valid-feedback">
+                        Looks good!
+                      </div>
+                    </div>
+                    <div class="col-md-4">
+                      <label for="validationCustomUsername"
+                             class="form-label">Username
+                      </label>
+                      <div class="input-group has-validation">
+                        <span class="input-group-text"
+                              id="inputGroupPrepend">@
+                        </span>
+                        <input type="text"
+                               class="form-control"
+                               id="validationCustomUsername"
+                               aria-describedby="inputGroupPrepend"
+                               required>
+                        <div class="invalid-feedback">
+                          Please choose a username.
+                        </div>
+                      </div>
+                    </div>
+                    <div class="col-md-6">
+                      <label for="validationCustom03"
+                             class="form-label">City
+                      </label>
+                      <input type="text"
+                             class="form-control"
+                             id="validationCustom03"
+                             required>
+                      <div class="invalid-feedback">
+                        Please provide a valid city.
+                      </div>
+                    </div>
+                    <div class="col-md-3">
+                      <label for="validationCustom04"
+                             class="form-label">State
+                      </label>
+                      <select class="form-select"
+                              id="validationCustom04"
+                              required>
+                        <option selected
+                                disabled
+                                value="">Choose...
+                        </option>
+                        <option>...</option>
+                      </select>
+                      <div class="invalid-feedback">
+                        Please select a valid state.
+                      </div>
+                    </div>
+                    <div class="col-md-3">
+                      <label for="validationCustom05"
+                             class="form-label">Zip
+                      </label>
+                      <input type="text"
+                             class="form-control"
+                             id="validationCustom05"
+                             required>
+                      <div class="invalid-feedback">
+                        Please provide a valid zip.
+                      </div>
+                    </div>
+                    <div class="col-12">
+                      <div class="form-check">
+                        <input class="form-check-input"
+                               type="checkbox"
+                               value=""
+                               id="invalidCheck"
+                               required>
+                        <label class="form-check-label"
+                               for="invalidCheck">
+                          Agree to terms and conditions
+                        </label>
+                        <div class="invalid-feedback">
+                          You must agree before submitting.
+                        </div>
+                      </div>
+                    </div>
+                    <div class="col-12">
+                      <button class="btn btn-primary"
+                              type="submit">Submit form
+                      </button>
+                    </div>
+                  </div>
+                </form>
+
+                <div>
+
+                  <hr>
+                  <h2 class="fs-xlarge">Example validation invalid</h2>
+
+                  <form class="was-validated">
+                    <div class="row g-1">
+                      <div class="col-12">
+                        <label for="validationTextarea"
+                               class="form-label">Textarea
+                        </label>
+                        <textarea class="form-control"
+                                  id="validationTextarea"
+                                  placeholder="Required example textarea"
+                                  required></textarea>
+                        <div class="invalid-feedback">
+                          Please enter a message in the textarea.
+                        </div>
+                      </div>
+                      <div class="col-12">
+                        <div class="form-check">
+                          <input type="checkbox"
+                                 class="form-check-input"
+                                 id="validationFormCheck1"
+                                 required>
+                          <label class="form-check-label"
+                                 for="validationFormCheck1">
+                            Check this checkbox
+                          </label>
+                          <div class="invalid-feedback">
+                            Example invalid feedback text
+                          </div>
+                        </div>
+                        <div class="form-check">
+                          <input type="radio"
+                                 class="form-check-input"
+                                 id="validationFormCheck2"
+                                 name="radio-stacked"
+                                 required>
+                          <label class="form-check-label"
+                                 for="validationFormCheck2">
+                            Toggle this radio
+                          </label>
+                        </div>
+                        <div class="form-check">
+                          <input type="radio"
+                                 class="form-check-input"
+                                 id="validationFormCheck3"
+                                 name="radio-stacked"
+                                 required>
+                          <label class="form-check-label"
+                                 for="validationFormCheck3">
+                            Or toggle this other radio
+                          </label>
+                          <div class="invalid-feedback">
+                            More example invalid feedback text
+                          </div>
+                        </div>
+                      </div>
+                      <div class="col-md-6">
+                        <select class="form-select"
+                                required
+                                aria-label="select example">
+                          <option value="">Open this select menu</option>
+                          <option value="1">One</option>
+                          <option value="2">Two</option>
+                          <option value="3">Three</option>
+                        </select>
+                        <div class="invalid-feedback">
+                          Example invalid select feedback
+                        </div>
+                      </div>
+                      <div class="col-md-6">
+                        <input type="file"
+                               class="form-control"
+                               aria-label="file example"
+                               required>
+                        <div class="invalid-feedback">
+                          Example invalid form file feedback
+                        </div>
+                      </div>
+                      <div class="col-md-4">
+                        <button class="btn btn-primary"
+                                type="submit"
+                                disabled>Submit form
+                        </button>
+                      </div>
+                    </div>
+
+                  </form>
+                </div>
+
+              </div>
+
+            </div>
+
+            <div id="livewire"
+                 class="bd-example">
+
+              <h2 class="h6">Livewire</h2>
+              <div class="vstack gap-1">
+
+                <form wire:submit.prevent="submit"
+                      novalidate>
+                  <div class="row gx-1 gy-1h">
+                    <div class="col-12">
+                      <div class="form-floating">
+                        <input wire:model.lazy="name"
+                               type="text"
+                               id="name"
+                               placeholder="Seu nome"
+                               required
+                               aria-label="nome"
+                               class="@error('name') is-invalid @enderror form-control">
+                        <label for="name"
+                               class="form-label">Seu nome
+                        </label>
+                        @error('name')
+                        <span class="invalid-feedback">
+                          {{ $message }}
+                        </span>
+                        @enderror
+                      </div>
+                    </div>
+                    <div class="col-12">
+                      <div class="form-floating">
+                        <input wire:model.lazy="email"
+                               type="email"
+                               id="email"
+                               placeholder="Seu e-mail"
+                               required
+                               aria-label="email"
+                               class="@error('email') is-invalid @enderror form-control">
+                        <label for="email"
+                               class="form-label">Seu e-mail
+                        </label>
+                        @error('email')
+                        <span class="invalid-feedback">
+                          {{ $message }}
+                        </span>
+                        @enderror
+                      </div>
+                    </div>
+                    <div class="col-12">
+                      <div class="form-floating">
+                        <input wire:model.lazy="phone"
+                               type="tel"
+                               id="phone"
+                               placeholder="Seu celular"
+                               required
+                               aria-label="telefone"
+                               class="@error('phone') is-invalid @enderror js-mask-phone form-control">
+                        <label for="phone"
+                               class="form-label">Seu celular
+                        </label>
+                        @error('phone')
+                        <span class="invalid-feedback">
+                          {{ $message }}
+                        </span>
+                        @enderror
+                      </div>
+                    </div>
+                    <div class="col-12 col-md-5">
+                      <div class="form-floating">
+                        <select wire:model.lazy="state"
+                                id="state"
+                                required
+                                aria-label="estado"
+                                class="@error('state') is-invalid @enderror form-select">
+                          <option value=""></option>
+                          @php $states = ['SP' => 'SP'] @endphp
+                          @foreach($states as $value => $text)
+                          <option value="{{ $value }}">{{ $text }}</option>
+                          @endforeach
+                        </select>
+                        <label for="state"
+                               class="form-label">Seu estado
+                        </label>
+                        @error('state')
+                        <span class="invalid-feedback">
+                          {{ $message }}
+                        </span>
+                        @enderror
+                      </div>
+                    </div>
+                    <div class="col-12 col-md-7">
+                      <div class="form-floating">
+                        <select wire:model.lazy="city"
+                                id="city"
+                                required
+                                aria-label="city"
+                                class="@error('city') is-invalid @enderror form-select">
+                          <option value=""></option>
+                          @php $cities = ['São Paulo' => 'São Paulo'] @endphp
+                          @foreach($cities as $value => $text)
+                          <option value="{{ $value }}">{{ $text }}</option>
+                          @endforeach
+                        </select>
+                        <label for="city"
+                               class="form-label">Sua cidade
+                        </label>
+                        @error('city')
+                        <span class="invalid-feedback">
+                          {{ $message }}
+                        </span>
+                        @enderror
+                      </div>
+                    </div>
+                    <div class="col-12">
+                      <div class="form-floating">
+                        <textarea wire:model.lazy="message"
+                                  id="message"
+                                  placeholder="Sua mensagem"
+                                  required
+                                  aria-label="mensagem"
+                                  rows="5"
+                                  class="@error('message') is-invalid @enderror form-control"></textarea>
+                        <label for="message"
+                               class="form-label">Sua mensagem
+                        </label>
+                        @error('message')
+                        <span class="invalid-feedback">
+                          {{ $message }}
+                        </span>
+                        @enderror
+                      </div>
+                    </div>
+                    <div class="col-12">
+                      <div class="form-check">
+                        <input wire:model.lazy="terms"
+                               type="checkbox"
+                               name="privacy-terms"
+                               required
+                               class="form-check-input @error('terms') is-invalid @enderror"
+                               id="privacy-terms">
+                        <label class="form-check-label text-reset"
+                               for="privacy-terms">
+                          <x-frontend::privacy-terms-message/>
+                        </label>
+
+                        @error('terms')
+                        <span class="d-block invalid-feedback">
+                          {{ $message }}
+                        </span>
+                        @enderror
+                      </div>
+                    </div>
+                    <div class="col-12">
+                      <button type="submit"
+                              wire:loading.attr="disabled"
+                              wire:target="submit"
+                              aria-label="enviar formulário"
+                              class="btn btn-secondary w-100">
+                        <span wire:loading.remove
+                              wire:target="submit">Enviar
+                        </span>
+                        <span wire:loading
+                              wire:target="submit">
+                          Enviando...
+                          <span class="spinner-grow spinner-grow-sm ms-0h"
+                                role="status"
+                                aria-hidden="true"></span>
+                        </span>
+                      </button>
+                    </div>
+                    <div class="mt-1 hstack gap-0hq fs-small justify-content-center align-items-center">
+                      <x-frontend-icon name="ic-shield"
+                                       class="ic-xs"/>
+                      Seus dados estão seguros!
+                    </div>
+                  </div>
+                </form>
+
+              </div>
+
+            </div>
+
+          </div>
+
+        </div>
+
+      </div>
+    </div>
+  </div>
+</main>
+@endsection
+
 @push('head')
-    <style>
-        iframe {
-            overflow: hidden;
-            border: none
-        }
-
-        @media (min-width: 768px) {
-            body > .navbar-transparent {
-                box-shadow: none
-            }
-
-            body > .navbar-transparent .navbar-nav > .open > a {
-                box-shadow: none
-            }
-        }
-
-        .navbar .fa, .navbar .fab {
-            margin-right: .25rem;
-            font-size: 1rem
-        }
-
-        #home .navbar, #help .navbar {
-            background: #375ee3;
-            background: linear-gradient(145deg, #375ee3 0%, #6543e0 80%);
-            transition: box-shadow 200ms ease-in
-        }
-
-        #home .navbar-brand .nav-link, #help .navbar-brand .nav-link {
-            display: inline-block;
-            margin-right: -30px
-        }
-
-        #home {
-            padding-top: 0
-        }
-
-        #home .btn {
-            padding: .5rem .9rem
-        }
-
-        .bs-docs-section:not(:first-child) {
-            margin-top: 4em
-        }
-
-        .bs-docs-section .page-header h1 {
-            padding: 2rem 0;
-            font-size: 3rem
-        }
-
-        .dropdown-menu.show[aria-labelledby="themes"] {
-            display: -ms-flexbox;
-            display: flex;
-            width: 420px;
-            -ms-flex-wrap: wrap;
-            flex-wrap: wrap
-        }
-
-        .dropdown-menu.show[aria-labelledby="themes"] .dropdown-item {
-            width: 33.333%
-        }
-
-        .dropdown-menu.show[aria-labelledby="themes"] .dropdown-item:first-child {
-            width: 100%
-        }
-
-        .bs-component {
-            position: relative
-        }
-
-        .bs-component + .bs-component {
-            margin-top: 1rem
-        }
-
-        .bs-component .card {
-            margin-bottom: 1rem
-        }
-
-        .bs-component .modal {
-            position: relative;
-            top: auto;
-            right: auto;
-            left: auto;
-            bottom: auto;
-            z-index: 1;
-            display: block
-        }
-
-        .bs-component .modal-dialog {
-            width: 90%
-        }
-
-        .bs-component .popover {
-            position: relative;
-            display: inline-block;
-            width: 220px;
-            margin: 20px
-        }
-
-        .source-button {
-            display: none;
-            position: absolute;
-            top: 0;
-            right: 0;
-            z-index: 100;
-            font-weight: 700
-        }
-
-        .source-button:hover {
-            cursor: pointer
-        }
-
-        .bs-component:hover .source-button {
-            display: block
-        }
-
-        #source-modal pre {
-            max-height: calc(100vh - 11rem)
-        }
-
-        .nav-tabs {
-            margin-bottom: 15px
-        }
-
-        .progress {
-            margin-bottom: 10px
-        }
-
-        #footer {
-            margin: 5em 0
-        }
-
-        #footer li {
-            float: left;
-            margin-right: 1.5em;
-            margin-bottom: 1.5em
-        }
-
-        #footer p {
-            clear: left;
-            margin-bottom: 0
-        }
-
-        .splash {
-            position: relative;
-            padding: 12em 0 6em;
-            background-color: #375ee3;
-            background-image: url("../img/splash.svg"), linear-gradient(145deg, #375ee3 0%, #6543e0 80%);
-            background-size: cover;
-            background-repeat: no-repeat;
-            color: #fff;
-            text-align: center
-        }
-
-        .splash .logo {
-            width: 160px
-        }
-
-        .splash h1 {
-            font-size: 3em;
-            color: #fff
-        }
-
-        .splash #social {
-            margin: 2em 0 3em
-        }
-
-        .splash .alert {
-            margin: 2em 0;
-            border: none
-        }
-
-        .splash .sponsor a {
-            color: #fff
-        }
-
-        .section-tout {
-            padding: 6em 0 1em;
-            border-bottom: 1px solid rgba(0, 0, 0, 0.05);
-            background-color: #eaf1f1;
-            text-align: center
-        }
-
-        .section-tout .icon {
-            display: -ms-flexbox;
-            display: flex;
-            -ms-flex-pack: center;
-            justify-content: center;
-            -ms-flex-align: center;
-            align-items: center;
-            width: 80px;
-            height: 80px;
-            margin: 0 auto 1rem;
-            background: #375ee3;
-            background: linear-gradient(145deg, #375ee3 0%, #6543e0 80%);
-            border-radius: 50%;
-            font-size: 2rem;
-            color: rgba(255, 255, 255, 0.9)
-        }
-
-        .section-tout p {
-            margin-bottom: 5em
-        }
-
-        .section-preview {
-            padding: 4em 0
-        }
-
-        .section-preview .preview {
-            margin-bottom: 4em;
-            background-color: #eaf1f1
-        }
-
-        .section-preview .preview .image {
-            position: relative
-        }
-
-        .section-preview .preview .image img {
-            width: 100%;
-            height: auto
-        }
-
-        .section-preview .preview .options {
-            padding: 2em;
-            border-top: 1px solid rgba(0, 0, 0, 0.125);
-            text-align: center
-        }
-
-        .section-preview .preview .options p {
-            margin-bottom: 2em
-        }
-
-        .section-preview .dropdown-menu {
-            text-align: left
-        }
-
-        .section-preview .lead {
-            margin-bottom: 2em
-        }
-
-        .sponsor a {
-            text-decoration: none
-        }
-
-        .sponsor #carbonads {
-            max-width: 240px;
-            margin: 0 auto
-        }
-
-        .sponsor .carbon-text {
-            display: block;
-            margin-top: 1em;
-            font-size: 12px
-        }
-
-        .sponsor .carbon-poweredby {
-            float: right;
-            margin-top: 1em;
-            font-size: 10px
-        }
-
-        @media (max-width: 767px) {
-            .splash {
-                padding-top: 8em
-            }
-
-            .splash .logo {
-                width: 100px
-            }
-
-            .splash h1 {
-                font-size: 2em
-            }
-
-            #banner {
-                margin-bottom: 2em;
-                text-align: center
-            }
-        }
-
-        @media (max-width: 991px) {
-            .navbar .fa {
-                display: none
-            }
-        }
-
-        @media (min-width: 992px) {
-            .navbar-transparent {
-                background: none !important;
-                box-shadow: none
-            }
-        }
-
-        code[class*="language-"],
-        pre[class*="language-"] {
-            user-select: all;
-            color: #f8f8f2;
-            background: none;
-            text-shadow: 0 1px rgba(0, 0, 0, 0.3);
-            font-family: Consolas, Monaco, 'Andale Mono', 'Ubuntu Mono', monospace;
-            font-size: 1em;
-            text-align: left;
-            white-space: pre;
-            word-spacing: normal;
-            word-break: normal;
-            word-wrap: normal;
-            line-height: 1.5;
-
-            -moz-tab-size: 4;
-            -o-tab-size: 4;
-            tab-size: 4;
-
-            -webkit-hyphens: none;
-            -moz-hyphens: none;
-            -ms-hyphens: none;
-            hyphens: none;
-        }
-
-        /* Code blocks */
-        pre[class*="language-"] {
-            padding: 1em;
-            margin: .5em 0;
-            overflow: auto;
-            border-radius: 0.3em;
-        }
-
-        :not(pre) > code[class*="language-"],
-        pre[class*="language-"] {
-            background: #272822;
-        }
-
-        /* Inline code */
-        :not(pre) > code[class*="language-"] {
-            padding: .1em;
-            border-radius: .3em;
-            white-space: normal;
-        }
-
-        .token.comment,
-        .token.prolog,
-        .token.doctype,
-        .token.cdata {
-            color: #8292a2;
-        }
-
-        .token.punctuation {
-            color: #f8f8f2;
-        }
-
-        .token.namespace {
-            opacity: .7;
-        }
-
-        .token.property,
-        .token.tag,
-        .token.constant,
-        .token.symbol,
-        .token.deleted {
-            color: #f92672;
-        }
-
-        .token.boolean,
-        .token.number {
-            color: #ae81ff;
-        }
-
-        .token.selector,
-        .token.attr-name,
-        .token.string,
-        .token.char,
-        .token.builtin,
-        .token.inserted {
-            color: #a6e22e;
-        }
-
-        .token.operator,
-        .token.entity,
-        .token.url,
-        .language-css .token.string,
-        .style .token.string,
-        .token.variable {
-            color: #f8f8f2;
-        }
-
-        .token.atrule,
-        .token.attr-value,
-        .token.function,
-        .token.class-name {
-            color: #e6db74;
-        }
-
-        .token.keyword {
-            color: #66d9ef;
-        }
-
-        .token.regex,
-        .token.important {
-            color: #fd971f;
-        }
-
-        .token.important,
-        .token.bold {
-            font-weight: bold;
-        }
-
-        .token.italic {
-            font-style: italic;
-        }
-
-        .token.entity {
-            cursor: help;
-        }
-    </style>
+<style>
+  .bd-example {
+    padding: 1rem;
+    border: 1px solid #dee2e6;
+    border-radius: .375rem;
+  }
+
+  .bd-example-code {
+    padding: 1rem;
+    border: 1px solid #dee2e6;
+    border-radius: .375rem;
+    background: #f8f9fa;
+  }
+</style>
 @endpush
 
-@section('content')
+@section('footer') @endsection
 
-    <main id="banner" class="container">
-
-        <!-- Navbar
-        ================================================== -->
-        <div class="bs-docs-section clearfix">
-            <div class="row">
-                <div class="col-lg-12">
-                    <div class="page-header">
-                        <h1 id="navbars">Navbars</h1>
-                    </div>
-
-                    <div class="bs-component">
-                        <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
-                            <div class="container-fluid">
-                                <a class="navbar-brand" href="#">Navbar</a>
-                                <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
-                                        data-bs-target="#navbarColor01" aria-controls="navbarColor01"
-                                        aria-expanded="false"
-                                        aria-label="Toggle navigation">
-                                    <span class="navbar-toggler-icon"></span>
-                                </button>
-
-                                <div class="collapse navbar-collapse" id="navbarColor01">
-                                    <ul class="navbar-nav me-auto">
-                                        <li class="nav-item">
-                                            <a class="nav-link active"
-                                               href="#">Home
-                                                <span class="visually-hidden">(current)</span>
-                                            </a>
-                                        </li>
-                                        <li class="nav-item">
-                                            <a class="nav-link" href="#">Features</a>
-                                        </li>
-                                        <li class="nav-item">
-                                            <a class="nav-link" href="#">Pricing</a>
-                                        </li>
-                                        <li class="nav-item">
-                                            <a class="nav-link" href="#">About</a>
-                                        </li>
-                                        <li class="nav-item dropdown">
-                                            <a class="nav-link dropdown-toggle" data-bs-toggle="dropdown" href="#"
-                                               role="button" aria-haspopup="true" aria-expanded="false">Dropdown</a>
-                                            <div class="dropdown-menu">
-                                                <a class="dropdown-item" href="#">Action</a>
-                                                <a class="dropdown-item" href="#">Another action</a>
-                                                <a class="dropdown-item" href="#">Something else here</a>
-                                                <div class="dropdown-divider"></div>
-                                                <a class="dropdown-item" href="#">Separated link</a>
-                                            </div>
-                                        </li>
-                                    </ul>
-                                    <form class="d-flex">
-                                        <input class="form-control me-sm-2" type="text" placeholder="Search">
-                                        <button class="btn btn-secondary my-2 my-sm-0" type="submit">Search</button>
-                                    </form>
-                                </div>
-                            </div>
-                        </nav>
-                    </div>
-
-                    <div class="bs-component">
-                        <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-                            <div class="container-fluid">
-                                <a class="navbar-brand" href="#">Navbar</a>
-                                <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
-                                        data-bs-target="#navbarColor02" aria-controls="navbarColor02"
-                                        aria-expanded="false"
-                                        aria-label="Toggle navigation">
-                                    <span class="navbar-toggler-icon"></span>
-                                </button>
-
-                                <div class="collapse navbar-collapse" id="navbarColor02">
-                                    <ul class="navbar-nav me-auto">
-                                        <li class="nav-item">
-                                            <a class="nav-link active"
-                                               href="#">Home
-                                                <span class="visually-hidden">(current)</span>
-                                            </a>
-                                        </li>
-                                        <li class="nav-item">
-                                            <a class="nav-link" href="#">Features</a>
-                                        </li>
-                                        <li class="nav-item">
-                                            <a class="nav-link" href="#">Pricing</a>
-                                        </li>
-                                        <li class="nav-item">
-                                            <a class="nav-link" href="#">About</a>
-                                        </li>
-                                        <li class="nav-item dropdown">
-                                            <a class="nav-link dropdown-toggle" data-bs-toggle="dropdown" href="#"
-                                               role="button" aria-haspopup="true" aria-expanded="false">Dropdown</a>
-                                            <div class="dropdown-menu">
-                                                <a class="dropdown-item" href="#">Action</a>
-                                                <a class="dropdown-item" href="#">Another action</a>
-                                                <a class="dropdown-item" href="#">Something else here</a>
-                                                <div class="dropdown-divider"></div>
-                                                <a class="dropdown-item" href="#">Separated link</a>
-                                            </div>
-                                        </li>
-                                    </ul>
-                                    <form class="d-flex">
-                                        <input class="form-control me-sm-2" type="text" placeholder="Search">
-                                        <button class="btn btn-secondary my-2 my-sm-0" type="submit">Search</button>
-                                    </form>
-                                </div>
-                            </div>
-                        </nav>
-                    </div>
-
-                    <div class="bs-component">
-                        <nav class="navbar navbar-expand-lg navbar-light bg-light">
-                            <div class="container-fluid">
-                                <a class="navbar-brand" href="#">Navbar</a>
-                                <button class="navbar-toggler" type="button" data-bs-toggle="collapse"
-                                        data-bs-target="#navbarColor03" aria-controls="navbarColor03"
-                                        aria-expanded="false"
-                                        aria-label="Toggle navigation">
-                                    <span class="navbar-toggler-icon"></span>
-                                </button>
-
-                                <div class="collapse navbar-collapse" id="navbarColor03">
-                                    <ul class="navbar-nav me-auto">
-                                        <li class="nav-item">
-                                            <a class="nav-link active"
-                                               href="#">Home
-                                                <span class="visually-hidden">(current)</span>
-                                            </a>
-                                        </li>
-                                        <li class="nav-item">
-                                            <a class="nav-link" href="#">Features</a>
-                                        </li>
-                                        <li class="nav-item">
-                                            <a class="nav-link" href="#">Pricing</a>
-                                        </li>
-                                        <li class="nav-item">
-                                            <a class="nav-link" href="#">About</a>
-                                        </li>
-                                        <li class="nav-item dropdown">
-                                            <a class="nav-link dropdown-toggle" data-bs-toggle="dropdown" href="#"
-                                               role="button" aria-haspopup="true" aria-expanded="false">Dropdown</a>
-                                            <div class="dropdown-menu">
-                                                <a class="dropdown-item" href="#">Action</a>
-                                                <a class="dropdown-item" href="#">Another action</a>
-                                                <a class="dropdown-item" href="#">Something else here</a>
-                                                <div class="dropdown-divider"></div>
-                                                <a class="dropdown-item" href="#">Separated link</a>
-                                            </div>
-                                        </li>
-                                    </ul>
-                                    <form class="d-flex">
-                                        <input class="form-control me-sm-2" type="text" placeholder="Search">
-                                        <button class="btn btn-secondary my-2 my-sm-0" type="submit">Search</button>
-                                    </form>
-                                </div>
-                            </div>
-                        </nav>
-                    </div>
-
-                </div>
-            </div>
-        </div>
-
-        <!-- Buttons
-        ================================================== -->
-        <div class="bs-docs-section">
-            <div class="page-header">
-                <div class="row">
-                    <div class="col-lg-12">
-                        <h1 id="buttons">Buttons</h1>
-                    </div>
-                </div>
-            </div>
-
-            <div class="row">
-                <div class="col-lg-7">
-                    <p class="bs-component">
-                        <button type="button" class="btn btn-primary">Primary</button>
-                        <button type="button" class="btn btn-secondary">Secondary</button>
-                        <button type="button" class="btn btn-success">Success</button>
-                        <button type="button" class="btn btn-info">Info</button>
-                        <button type="button" class="btn btn-warning">Warning</button>
-                        <button type="button"
-                                class="btn btn-danger">Danger
-                        </button>
-                        <button type="button"
-                                class="btn btn-light">Light
-                        </button>
-                        <button type="button"
-                                class="btn btn-dark">Dark
-                        </button>
-                        <button type="button"
-                                class="btn btn-link">Link
-                        </button>
-                    </p>
-
-                    <p class="bs-component">
-                        <button type="button" class="btn btn-primary disabled">Primary</button>
-                        <button type="button" class="btn btn-secondary disabled">Secondary</button>
-                        <button type="button" class="btn btn-success disabled">Success</button>
-                        <button type="button" class="btn btn-info disabled">Info</button>
-                        <button type="button" class="btn btn-warning disabled">Warning</button>
-                        <button type="button"
-                                class="btn btn-danger disabled">Danger
-                        </button>
-                        <button type="button"
-                                class="btn btn-light disabled">Light
-                        </button>
-                        <button type="button"
-                                class="btn btn-dark disabled">Dark
-                        </button>
-                        <button type="button" class="btn btn-link disabled">Link</button>
-                    </p>
-
-                    <p class="bs-component">
-                        <button type="button" class="btn btn-outline-primary">Primary</button>
-                        <button type="button" class="btn btn-outline-secondary">Secondary</button>
-                        <button type="button" class="btn btn-outline-success">Success</button>
-                        <button type="button" class="btn btn-outline-info">Info</button>
-                        <button type="button" class="btn btn-outline-warning">Warning</button>
-                        <button type="button"
-                                class="btn btn-outline-danger">Danger
-                        </button>
-                        <button type="button"
-                                class="btn btn-outline-light">Light
-                        </button>
-                        <button type="button"
-                                class="btn btn-outline-dark">Dark
-                        </button>
-                    </p>
-
-                    <div class="bs-component">
-                        <div class="btn-group" role="group" aria-label="Button group with nested dropdown">
-                            <button type="button" class="btn btn-primary">Primary</button>
-                            <div class="btn-group" role="group">
-                                <button id="btnGroupDrop1" type="button" class="btn btn-primary dropdown-toggle"
-                                        data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false"></button>
-                                <div class="dropdown-menu" aria-labelledby="btnGroupDrop1">
-                                    <a class="dropdown-item" href="#">Dropdown link</a>
-                                    <a class="dropdown-item" href="#">Dropdown link</a>
-                                </div>
-                            </div>
-                        </div>
-
-                        <div class="btn-group" role="group" aria-label="Button group with nested dropdown">
-                            <button type="button" class="btn btn-success">Success</button>
-                            <div class="btn-group" role="group">
-                                <button id="btnGroupDrop2" type="button" class="btn btn-success dropdown-toggle"
-                                        data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false"></button>
-                                <div class="dropdown-menu" aria-labelledby="btnGroupDrop2">
-                                    <a class="dropdown-item" href="#">Dropdown link</a>
-                                    <a class="dropdown-item" href="#">Dropdown link</a>
-                                </div>
-                            </div>
-                        </div>
-
-                        <div class="btn-group" role="group" aria-label="Button group with nested dropdown">
-                            <button type="button" class="btn btn-info">Info</button>
-                            <div class="btn-group" role="group">
-                                <button id="btnGroupDrop3" type="button" class="btn btn-info dropdown-toggle"
-                                        data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false"></button>
-                                <div class="dropdown-menu" aria-labelledby="btnGroupDrop3">
-                                    <a class="dropdown-item" href="#">Dropdown link</a>
-                                    <a class="dropdown-item" href="#">Dropdown link</a>
-                                </div>
-                            </div>
-                        </div>
-
-                        <div class="btn-group" role="group" aria-label="Button group with nested dropdown">
-                            <button type="button" class="btn btn-danger">Danger</button>
-                            <div class="btn-group" role="group">
-                                <button id="btnGroupDrop4" type="button" class="btn btn-danger dropdown-toggle"
-                                        data-bs-toggle="dropdown" aria-haspopup="true" aria-expanded="false"></button>
-                                <div class="dropdown-menu" aria-labelledby="btnGroupDrop4">
-                                    <a class="dropdown-item" href="#">Dropdown link</a>
-                                    <a class="dropdown-item" href="#">Dropdown link</a>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-
-                    <div class="bs-component">
-                        <button type="button" class="btn btn-primary btn-lg">Large button</button>
-                        <button type="button" class="btn btn-primary">Default button</button>
-                        <button type="button" class="btn btn-primary btn-sm">Small button</button>
-                    </div>
-                </div>
-                <div class="col-lg-5">
-                    <div class="bs-component">
-                        <div class="d-grid gap-2">
-                            <button class="btn btn-lg btn-primary" type="button">Block button</button>
-                            <button class="btn btn-lg btn-primary" type="button">Block button</button>
-                        </div>
-                    </div>
-
-                    <div class="bs-component mb-1">
-                        <div class="btn-group"
-                             role="group"
-                             aria-label="Basic checkbox toggle button group">
-                            <input type="checkbox"
-                                   class="btn-check"
-                                   id="btncheck1"
-                                   checked
-                                   autocomplete="off">
-                            <label class="btn btn-primary"
-                                   for="btncheck1">Checkbox 1
-                            </label>
-                            <input type="checkbox"
-                                   class="btn-check"
-                                   id="btncheck2"
-                                   autocomplete="off">
-                            <label class="btn btn-primary"
-                                   for="btncheck2">Checkbox 2
-                            </label>
-                            <input type="checkbox"
-                                   class="btn-check"
-                                   id="btncheck3"
-                                   autocomplete="off">
-                            <label class="btn btn-primary"
-                                   for="btncheck3">Checkbox 3
-                            </label>
-                        </div>
-                    </div>
-
-                    <div class="bs-component mb-1">
-                        <div class="btn-group" role="group" aria-label="Basic radio toggle button group">
-                            <input type="radio" class="btn-check" name="btnradio" id="btnradio1" autocomplete="off"
-                                   checked>
-                            <label class="btn btn-outline-primary" for="btnradio1">Radio 1</label>
-                            <input type="radio" class="btn-check" name="btnradio" id="btnradio2" autocomplete="off"
-                                   checked>
-                            <label class="btn btn-outline-primary" for="btnradio2">Radio 2</label>
-                            <input type="radio" class="btn-check" name="btnradio" id="btnradio3" autocomplete="off"
-                                   checked>
-                            <label class="btn btn-outline-primary" for="btnradio3">Radio 3</label>
-                        </div>
-                    </div>
-
-                    <div class="bs-component">
-                        <div class="btn-group-vertical">
-                            <button type="button" class="btn btn-primary">Button</button>
-                            <button type="button" class="btn btn-primary">Button</button>
-                            <button type="button" class="btn btn-primary">Button</button>
-                            <button type="button" class="btn btn-primary">Button</button>
-                            <button type="button" class="btn btn-primary">Button</button>
-                            <button type="button" class="btn btn-primary">Button</button>
-                        </div>
-                    </div>
-
-                    <div class="bs-component mb-1">
-                        <div class="btn-group" role="group" aria-label="Basic example">
-                            <button type="button" class="btn btn-secondary">Left</button>
-                            <button type="button" class="btn btn-secondary">Middle</button>
-                            <button type="button" class="btn btn-secondary">Right</button>
-                        </div>
-                    </div>
-
-                    <div class="bs-component mb-1">
-                        <div class="btn-toolbar"
-                             role="toolbar"
-                             aria-label="Toolbar with button groups">
-                            <div class="btn-group me-2"
-                                 role="group"
-                                 aria-label="First group">
-                                <button type="button"
-                                        class="btn btn-secondary">
-                                    1
-                                </button>
-                                <button type="button"
-                                        class="btn btn-secondary">
-                                    2
-                                </button>
-                                <button type="button"
-                                        class="btn btn-secondary">
-                                    3
-                                </button>
-                                <button type="button"
-                                        class="btn btn-secondary">
-                                    4
-                                </button>
-                            </div>
-                            <div class="btn-group me-2"
-                                 role="group"
-                                 aria-label="Second group">
-                                <button type="button"
-                                        class="btn btn-secondary">
-                                    5
-                                </button>
-                                <button type="button"
-                                        class="btn btn-secondary">
-                                    6
-                                </button>
-                                <button type="button"
-                                        class="btn btn-secondary">
-                                    7
-                                </button>
-                            </div>
-                            <div class="btn-group"
-                                 role="group"
-                                 aria-label="Third group">
-                                <button type="button"
-                                        class="btn btn-secondary">
-                                    8
-                                </button>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <!--Share Buttons-->
-        <div class="bs-component">
-            <h1>
-                Share buttons
-            </h1>
-            <div class="mt-1">
-                <a role="button"
-                   data-fmd-share-btn-fbidentification="123456789123456"
-                   class="btn btn-primary js-btn-share">
-                    share
-                </a>
-            </div>
-        </div>
-
-        <!-- Typography
-        ================================================== -->
-        <div class="bs-docs-section">
-            <div class="row">
-                <div class="col-lg-12">
-                    <div class="page-header">
-                        <h1 id="typography">Typography</h1>
-                    </div>
-                </div>
-            </div>
-
-            <!-- Headings -->
-
-            <div class="row">
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <h1>Heading 1</h1>
-                        <h2>Heading 2</h2>
-                        <h3>Heading 3</h3>
-                        <h4>Heading 4</h4>
-                        <h5>Heading 5</h5>
-                        <h6>Heading 6</h6>
-                        <h3>
-                            Heading
-                            <small class="text-muted">with muted text</small>
-                        </h3>
-                        <p class="lead">Vivamus sagittis lacus vel augue laoreet rutrum faucibus dolor auctor.</p>
-                    </div>
-                </div>
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <h2>Example body text</h2>
-                        <p>Nullam quis risus eget <a href="#">urna mollis ornare</a> vel eu leo. Cum sociis natoque
-                            penatibus et magnis dis parturient montes, nascetur ridiculus mus. Nullam id dolor id nibh
-                            ultricies vehicula.</p>
-                        <p><small>This line of text is meant to be treated as fine print.</small></p>
-                        <p>The following is <strong>rendered as bold text</strong>.</p>
-                        <p>The following is <em>rendered as italicized text</em>.</p>
-                        <p>An abbreviation of the word attribute is <abbr title="attribute">attr</abbr>.</p>
-                    </div>
-                </div>
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <h2>Emphasis classes</h2>
-                        <p class="text-muted">Fusce dapibus, tellus ac cursus commodo, tortor mauris nibh.</p>
-                        <p class="text-primary">Nullam id dolor id nibh ultricies vehicula ut id elit.</p>
-                        <p class="text-secondary">Pellentesque ornare sem lacinia quam venenatis vestibulum.</p>
-                        <p class="text-warning">Etiam porta sem malesuada magna mollis euismod.</p>
-                        <p class="text-danger">Donec ullamcorper nulla non metus auctor fringilla.</p>
-                        <p class="text-success">Duis mollis, est non commodo luctus, nisi erat porttitor ligula.</p>
-                        <p class="text-info">Maecenas sed diam eget risus varius blandit sit amet non magna.</p>
-                    </div>
-                </div>
-            </div>
-
-            <!-- Blockquotes -->
-
-            <div class="row">
-                <div class="col-lg-12">
-                    <h2 id="type-blockquotes">Blockquotes</h2>
-                </div>
-            </div>
-            <div class="row">
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <figure>
-                            <blockquote class="blockquote">
-                                <p class="mb-0">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer posuere
-                                    erat a ante.</p>
-
-                            </blockquote>
-                            <figcaption class="blockquote-footer">
-                                Someone famous in <cite title="Source Title">Source Title</cite>
-                            </figcaption>
-                        </figure>
-                    </div>
-                </div>
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <figure class="text-center">
-                            <blockquote class="blockquote">
-                                <p class="mb-0">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer posuere
-                                    erat a ante.</p>
-
-                            </blockquote>
-                            <figcaption class="blockquote-footer">
-                                Someone famous in <cite title="Source Title">Source Title</cite>
-                            </figcaption>
-                        </figure>
-                    </div>
-                </div>
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <figure class="text-end">
-                            <blockquote class="blockquote">
-                                <p class="mb-0">Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer posuere
-                                    erat a ante.</p>
-
-                            </blockquote>
-                            <figcaption class="blockquote-footer">
-                                Someone famous in <cite title="Source Title">Source Title</cite>
-                            </figcaption>
-                        </figure>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <!-- Tables
-        ================================================== -->
-        <div class="bs-docs-section">
-            <div class="row">
-                <div class="col-lg-12">
-                    <div class="page-header">
-                        <h1 id="tables">Tables</h1>
-                    </div>
-
-                    <div class="bs-component">
-                        <table class="table table-hover">
-                            <thead>
-                            <tr>
-                                <th scope="col">Type</th>
-                                <th scope="col">Column heading</th>
-                                <th scope="col">Column heading</th>
-                                <th scope="col">Column heading</th>
-                            </tr>
-                            </thead>
-                            <tbody>
-                            <tr class="table-active">
-                                <th scope="row">Active</th>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                            </tr>
-                            <tr>
-                                <th scope="row">Default</th>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                            </tr>
-                            <tr class="table-primary">
-                                <th scope="row">Primary</th>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                            </tr>
-                            <tr class="table-secondary">
-                                <th scope="row">Secondary</th>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                            </tr>
-                            <tr class="table-success">
-                                <th scope="row">Success</th>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                            </tr>
-                            <tr class="table-danger">
-                                <th scope="row">Danger</th>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                            </tr>
-                            <tr class="table-warning">
-                                <th scope="row">Warning</th>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                            </tr>
-                            <tr class="table-info">
-                                <th scope="row">Info</th>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                            </tr>
-                            <tr class="table-light">
-                                <th scope="row">Light</th>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                            </tr>
-                            <tr class="table-dark">
-                                <th scope="row">Dark</th>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                                <td>Column content</td>
-                            </tr>
-                            </tbody>
-                        </table>
-                    </div><!-- /example -->
-                </div>
-            </div>
-        </div>
-
-        <!-- Forms
-        ================================================== -->
-        <div class="bs-docs-section">
-            <div class="row">
-                <div class="col-lg-12">
-                    <div class="page-header">
-                        <h1 id="forms">Forms</h1>
-                    </div>
-                </div>
-            </div>
-
-            <div class="row">
-                <div class="col-lg-6">
-                    <div class="bs-component">
-                        <form>
-                            <fieldset>
-                                <legend>Legend</legend>
-                                <div class="form-group row">
-                                    <label for="staticEmail" class="col-sm-2 col-form-label">Email</label>
-                                    <div class="col-sm-10">
-                                        <input type="text" readonly class="form-control-plaintext" id="staticEmail"
-                                               value="email@example.com">
-                                    </div>
-                                </div>
-                                <div class="form-group">
-                                    <label for="exampleInputEmail1"
-                                           class="form-label mt-1h">Email address
-                                    </label>
-                                    <input type="email" class="form-control" id="exampleInputEmail1"
-                                           aria-describedby="emailHelp" placeholder="Enter email">
-                                    <small id="emailHelp" class="form-text text-muted">We'll never share your email with
-                                        anyone else.</small>
-                                </div>
-                                <div class="form-group">
-                                    <label for="exampleInputPassword1"
-                                           class="form-label mt-1h">Password
-                                    </label>
-                                    <input type="password" class="form-control" id="exampleInputPassword1"
-                                           placeholder="Password">
-                                </div>
-                                <div class="form-group">
-                                    <label for="exampleSelect1"
-                                           class="form-label mt-1h">Example select
-                                    </label>
-                                    <select class="form-select" id="exampleSelect1">
-                                        <option>1</option>
-                                        <option>2</option>
-                                        <option>3</option>
-                                        <option>4</option>
-                                        <option>5</option>
-                                    </select>
-                                </div>
-                                <div class="form-group">
-                                    <label for="exampleSelect2"
-                                           class="form-label mt-1h">Example multiple select
-                                    </label>
-                                    <select multiple class="form-select" id="exampleSelect2">
-                                        <option>1</option>
-                                        <option>2</option>
-                                        <option>3</option>
-                                        <option>4</option>
-                                        <option>5</option>
-                                    </select>
-                                </div>
-                                <div class="form-group">
-                                    <label for="exampleTextarea"
-                                           class="form-label mt-1h">Example textarea
-                                    </label>
-                                    <textarea class="form-control" id="exampleTextarea" rows="3"></textarea>
-                                </div>
-                                <div class="form-group">
-                                    <label for="formFile"
-                                           class="form-label mt-1h">Default file input example
-                                    </label>
-                                    <input class="form-control" type="file" id="formFile">
-                                </div>
-                                <fieldset class="form-group">
-                                    <legend class="mt-1h">Radio buttons</legend>
-                                    <div class="form-check">
-                                        <label class="form-check-label">
-                                            <input type="radio" class="form-check-input" name="optionsRadios"
-                                                   id="optionsRadios1" value="option1" checked>
-                                            Option one is this and that&mdash;be sure to include why it's great
-                                        </label>
-                                    </div>
-                                    <div class="form-check">
-                                        <label class="form-check-label">
-                                            <input type="radio" class="form-check-input" name="optionsRadios"
-                                                   id="optionsRadios2" value="option2">
-                                            Option two can be something else and selecting it will deselect option one
-                                        </label>
-                                    </div>
-                                    <div class="form-check disabled">
-                                        <label class="form-check-label">
-                                            <input type="radio" class="form-check-input" name="optionsRadios"
-                                                   id="optionsRadios3" value="option3" disabled>
-                                            Option three is disabled
-                                        </label>
-                                    </div>
-                                </fieldset>
-                                <fieldset class="form-group">
-                                    <legend class="mt-1h">Checkboxes</legend>
-                                    <div class="form-check">
-                                        <input class="form-check-input" type="checkbox" value="" id="flexCheckDefault">
-                                        <label class="form-check-label" for="flexCheckDefault">
-                                            Default checkbox
-                                        </label>
-                                    </div>
-                                    <div class="form-check">
-                                        <input class="form-check-input" type="checkbox" value="" id="flexCheckChecked"
-                                               checked>
-                                        <label class="form-check-label" for="flexCheckChecked">
-                                            Checked checkbox
-                                        </label>
-                                    </div>
-                                </fieldset>
-                                <fieldset>
-                                    <legend class="mt-1h">Switches</legend>
-                                    <div class="form-check form-switch">
-                                        <input class="form-check-input" type="checkbox" id="flexSwitchCheckDefault">
-                                        <label class="form-check-label" for="flexSwitchCheckDefault">Default switch
-                                            checkbox
-                                            input</label>
-                                    </div>
-                                    <div class="form-check form-switch">
-                                        <input class="form-check-input" type="checkbox" id="flexSwitchCheckChecked"
-                                               checked>
-                                        <label class="form-check-label" for="flexSwitchCheckChecked">Checked switch
-                                            checkbox
-                                            input</label>
-                                    </div>
-                                </fieldset>
-                                <fieldset class="form-group">
-                                    <legend class="mt-1h">Ranges</legend>
-                                    <label for="customRange1" class="form-label">Example range</label>
-                                    <input type="range" class="form-range" id="customRange1">
-                                    <label for="disabledRange" class="form-label">Disabled range</label>
-                                    <input type="range" class="form-range" id="disabledRange" disabled>
-                                    <label for="customRange3" class="form-label">Example range</label>
-                                    <input type="range" class="form-range" min="0" max="5" step="0.5" id="customRange3">
-                                </fieldset>
-                                <button type="submit" class="btn btn-primary">Submit</button>
-                            </fieldset>
-                        </form>
-                    </div>
-                </div>
-                <div class="col-lg-4 offset-lg-1">
-                    <form class="bs-component">
-                        <div class="form-group">
-                            <fieldset disabled>
-                                <label class="form-label" for="disabledInput">Disabled input</label>
-                                <input class="form-control" id="disabledInput" type="text"
-                                       placeholder="Disabled input here..." disabled>
-                            </fieldset>
-                        </div>
-
-                        <div class="form-group">
-                            <fieldset>
-                                <label class="form-label mt-1h"
-                                       for="readOnlyInput">Readonly input
-                                </label>
-                                <input class="form-control" id="readOnlyInput" type="text"
-                                       placeholder="Readonly input here..." readonly>
-                            </fieldset>
-                        </div>
-
-                        <div class="form-group has-success">
-                            <label class="form-label mt-1h"
-                                   for="inputValid">Valid input
-                            </label>
-                            <input type="text" value="correct value" class="form-control is-valid" id="inputValid">
-                            <div class="valid-feedback">Success! You've done it.</div>
-                        </div>
-
-                        <div class="form-group has-danger">
-                            <label class="form-label mt-1h"
-                                   for="inputInvalid">Invalid input
-                            </label>
-                            <input type="text" value="wrong value" class="form-control is-invalid" id="inputInvalid">
-                            <div class="invalid-feedback">Sorry, that username's taken. Try another?</div>
-                        </div>
-
-                        <div class="form-group">
-                            <label class="col-form-label col-form-label-lg mt-1h"
-                                   for="inputLarge">Large input
-                            </label>
-                            <input class="form-control form-control-lg" type="text" placeholder=".form-control-lg"
-                                   id="inputLarge">
-                        </div>
-
-                        <div class="form-group">
-                            <label class="col-form-label mt-1h"
-                                   for="inputDefault">Default input
-                            </label>
-                            <input type="text" class="form-control" placeholder="Default input" id="inputDefault">
-                        </div>
-
-                        <div class="form-group">
-                            <label class="col-form-label col-form-label-sm mt-1h"
-                                   for="inputSmall">Small input
-                            </label>
-                            <input class="form-control form-control-sm" type="text" placeholder=".form-control-sm"
-                                   id="inputSmall">
-                        </div>
-
-                        <div class="form-group">
-                            <label class="form-label mt-1h">Input addons</label>
-                            <div class="form-group">
-                                <div class="input-group mb-1">
-                                    <span class="input-group-text">$</span>
-                                    <input type="text" class="form-control" aria-label="Amount (to the nearest dollar)">
-                                    <span class="input-group-text">.00</span>
-                                </div>
-                                <div class="input-group mb-1">
-                                    <input type="text"
-                                           class="form-control"
-                                           placeholder="Recipient's username"
-                                           aria-label="Recipient's username"
-                                           aria-describedby="button-addon2">
-                                    <button class="btn btn-primary"
-                                            type="button"
-                                            id="button-addon2">Button
-                                    </button>
-                                </div>
-                            </div>
-                        </div>
-
-                        <div class="form-group">
-                            <label class="form-label mt-1h">Floating labels</label>
-                            <div class="form-floating mb-1">
-                                <input type="email" class="form-control" id="floatingInput"
-                                       placeholder="name@example.com">
-                                <label for="floatingInput">Email address</label>
-                            </div>
-                            <div class="form-floating">
-                                <input type="password" class="form-control" id="floatingPassword"
-                                       placeholder="Password">
-                                <label for="floatingPassword">Password</label>
-                            </div>
-                        </div>
-                    </form>
-
-                </div>
-            </div>
-        </div>
-
-        <!-- Livewire example
-          ================================================== -->
-        <div class="bs-docs-section">
-            <div class="page-header">
-                <h1 id="livewireExample">Livewire Example</h1>
-            </div>
-            <livewire:frontend::contact/>
-        </div>
-
-        <!-- Navs
-        ================================================== -->
-        <div class="bs-docs-section">
-            <div class="row">
-                <div class="col-lg-12">
-                    <div class="page-header">
-                        <h1 id="navs">Navs</h1>
-                    </div>
-                </div>
-            </div>
-
-            <div class="row mb-5">
-                <div class="col-lg-6">
-                    <h2 id="nav-tabs">Tabs</h2>
-                    <div class="bs-component">
-                        <ul class="nav nav-tabs">
-                            <li class="nav-item">
-                                <a class="nav-link active" data-bs-toggle="tab" href="#home">Home</a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" data-bs-toggle="tab" href="#profile">Profile</a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link disabled" href="#">Disabled</a>
-                            </li>
-                            <li class="nav-item dropdown">
-                                <a class="nav-link dropdown-toggle" data-bs-toggle="dropdown" href="#" role="button"
-                                   aria-haspopup="true" aria-expanded="false">Dropdown</a>
-                                <div class="dropdown-menu">
-                                    <a class="dropdown-item" href="#">Action</a>
-                                    <a class="dropdown-item" href="#">Another action</a>
-                                    <a class="dropdown-item" href="#">Something else here</a>
-                                    <div class="dropdown-divider"></div>
-                                    <a class="dropdown-item" href="#">Separated link</a>
-                                </div>
-                            </li>
-                        </ul>
-                        <div id="myTabContent" class="tab-content">
-                            <div class="tab-pane fade show active" id="home">
-                                <p>Raw denim you probably haven't heard of them jean shorts Austin. Nesciunt tofu
-                                    stumptown
-                                    aliqua, retro synth master cleanse. Mustache cliche tempor, williamsburg carles
-                                    vegan
-                                    helvetica. Reprehenderit butcher retro keffiyeh dreamcatcher synth. Cosby sweater eu
-                                    banh mi, qui irure terry richardson ex squid. Aliquip placeat salvia cillum iphone.
-                                    Seitan aliquip quis cardigan american apparel, butcher voluptate nisi qui.</p>
-                            </div>
-                            <div class="tab-pane fade" id="profile">
-                                <p>Food truck fixie locavore, accusamus mcsweeney's marfa nulla single-origin coffee
-                                    squid.
-                                    Exercitation +1 labore velit, blog sartorial PBR leggings next level wes anderson
-                                    artisan four loko farm-to-table craft beer twee. Qui photo booth letterpress,
-                                    commodo
-                                    enim craft beer mlkshk aliquip jean shorts ullamco ad vinyl cillum PBR. Homo nostrud
-                                    organic, assumenda labore aesthetic magna delectus mollit.</p>
-                            </div>
-                            <div class="tab-pane fade" id="dropdown1">
-                                <p>Etsy mixtape wayfarers, ethical wes anderson tofu before they sold out mcsweeney's
-                                    organic lomo retro fanny pack lo-fi farm-to-table readymade. Messenger bag gentrify
-                                    pitchfork tattooed craft beer, iphone skateboard locavore carles etsy salvia banksy
-                                    hoodie helvetica. DIY synth PBR banksy irony. Leggings gentrify squid 8-bit cred
-                                    pitchfork.</p>
-                            </div>
-                            <div class="tab-pane fade" id="dropdown2">
-                                <p>Trust fund seitan letterpress, keytar raw denim keffiyeh etsy art party before they
-                                    sold
-                                    out master cleanse gluten-free squid scenester freegan cosby sweater. Fanny pack
-                                    portland seitan DIY, art party locavore wolf cliche high life echo park Austin. Cred
-                                    vinyl keffiyeh DIY salvia PBR, banh mi before they sold out farm-to-table VHS viral
-                                    locavore cosby sweater.</p>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-
-                <div class="col-lg-6">
-                    <h2 id="nav-pills">Pills</h2>
-                    <div class="bs-component">
-                        <ul class="nav nav-pills">
-                            <li class="nav-item">
-                                <a class="nav-link active" href="#">Active</a>
-                            </li>
-                            <li class="nav-item dropdown">
-                                <a class="nav-link dropdown-toggle" data-bs-toggle="dropdown" href="#" role="button"
-                                   aria-haspopup="true" aria-expanded="false">Dropdown</a>
-                                <div class="dropdown-menu">
-                                    <a class="dropdown-item" href="#">Action</a>
-                                    <a class="dropdown-item" href="#">Another action</a>
-                                    <a class="dropdown-item" href="#">Something else here</a>
-                                    <div class="dropdown-divider"></div>
-                                    <a class="dropdown-item" href="#">Separated link</a>
-                                </div>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" href="#">Link</a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link disabled" href="#">Disabled</a>
-                            </li>
-                        </ul>
-                    </div>
-                    <br>
-                    <div class="bs-component">
-                        <ul class="nav nav-pills flex-column">
-                            <li class="nav-item">
-                                <a class="nav-link active" href="#">Active</a>
-                            </li>
-                            <li class="nav-item dropdown">
-                                <a class="nav-link dropdown-toggle" data-bs-toggle="dropdown" href="#" role="button"
-                                   aria-haspopup="true" aria-expanded="false">Dropdown</a>
-                                <div class="dropdown-menu">
-                                    <a class="dropdown-item" href="#">Action</a>
-                                    <a class="dropdown-item" href="#">Another action</a>
-                                    <a class="dropdown-item" href="#">Something else here</a>
-                                    <div class="dropdown-divider"></div>
-                                    <a class="dropdown-item" href="#">Separated link</a>
-                                </div>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link" href="#">Link</a>
-                            </li>
-                            <li class="nav-item">
-                                <a class="nav-link disabled" href="#">Disabled</a>
-                            </li>
-                        </ul>
-                    </div>
-                </div>
-            </div>
-
-            <div class="row">
-                <div class="col-lg-6">
-                    <h2 id="nav-breadcrumbs">Breadcrumbs</h2>
-                    <div class="bs-component">
-                        <ol class="breadcrumb">
-                            <li class="breadcrumb-item active">Home</li>
-                        </ol>
-                        <ol class="breadcrumb">
-                            <li class="breadcrumb-item"><a href="#">Home</a></li>
-                            <li class="breadcrumb-item active">Library</li>
-                        </ol>
-                        <ol class="breadcrumb">
-                            <li class="breadcrumb-item"><a href="#">Home</a></li>
-                            <li class="breadcrumb-item"><a href="#">Library</a></li>
-                            <li class="breadcrumb-item active">Data</li>
-                        </ol>
-                    </div>
-                </div>
-
-                <div class="col-lg-6">
-                    <h2 id="pagination">Pagination</h2>
-                    <div class="bs-component">
-                        <div>
-                            <ul class="pagination">
-                                <li class="page-item disabled">
-                                    <a class="page-link" href="#">&laquo;</a>
-                                </li>
-                                <li class="page-item active">
-                                    <a class="page-link" href="#">1</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">2</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">3</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">4</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">5</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">&raquo;</a>
-                                </li>
-                            </ul>
-                        </div>
-
-                        <div>
-                            <ul class="pagination pagination-lg">
-                                <li class="page-item disabled">
-                                    <a class="page-link" href="#">&laquo;</a>
-                                </li>
-                                <li class="page-item active">
-                                    <a class="page-link" href="#">1</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">2</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">3</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">4</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">5</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">&raquo;</a>
-                                </li>
-                            </ul>
-                        </div>
-
-                        <div>
-                            <ul class="pagination pagination-sm">
-                                <li class="page-item disabled">
-                                    <a class="page-link" href="#">&laquo;</a>
-                                </li>
-                                <li class="page-item active">
-                                    <a class="page-link" href="#">1</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">2</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">3</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">4</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">5</a>
-                                </li>
-                                <li class="page-item">
-                                    <a class="page-link" href="#">&raquo;</a>
-                                </li>
-                            </ul>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <!-- Indicators
-        ================================================== -->
-        <div class="bs-docs-section">
-            <div class="row">
-                <div class="col-lg-12">
-                    <div class="page-header">
-                        <h1 id="indicators">Indicators</h1>
-                    </div>
-                </div>
-            </div>
-
-            <div class="row">
-                <div class="col-lg-12">
-                    <h2>Alerts</h2>
-                    <div class="bs-component">
-                        <div class="alert alert-dismissible alert-warning">
-                            <button type="button" class="btn-close" data-bs-dismiss="alert"></button>
-                            <h4 class="alert-heading">Warning!</h4>
-                            <p class="mb-0">Best check yo self, you're not looking too good. Nulla vitae elit libero, a
-                                pharetra augue. Praesent commodo cursus magna, <a href="#" class="alert-link">vel
-                                    scelerisque nisl consectetur et</a>.</p>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="row">
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <div class="alert alert-dismissible alert-danger">
-                            <button type="button" class="btn-close" data-bs-dismiss="alert"></button>
-                            <strong>Oh snap!</strong> <a href="#" class="alert-link">Change a few things up</a> and try
-                            submitting again.
-                        </div>
-                    </div>
-                </div>
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <div class="alert alert-dismissible alert-success">
-                            <button type="button" class="btn-close" data-bs-dismiss="alert"></button>
-                            <strong>Well done!</strong> You successfully read <a href="#" class="alert-link">this
-                                important
-                                alert message</a>.
-                        </div>
-                    </div>
-                </div>
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <div class="alert alert-dismissible alert-info">
-                            <button type="button" class="btn-close" data-bs-dismiss="alert"></button>
-                            <strong>Heads up!</strong> This <a href="#" class="alert-link">alert needs your
-                                attention</a>,
-                            but it's not super important.
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="row">
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <div class="alert alert-dismissible alert-primary">
-                            <button type="button" class="btn-close" data-bs-dismiss="alert"></button>
-                            <strong>Oh snap!</strong> <a href="#" class="alert-link">Change a few things up</a> and try
-                            submitting again.
-                        </div>
-                    </div>
-                </div>
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <div class="alert alert-dismissible alert-secondary">
-                            <button type="button" class="btn-close" data-bs-dismiss="alert"></button>
-                            <strong>Well done!</strong> You successfully read <a href="#" class="alert-link">this
-                                important
-                                alert message</a>.
-                        </div>
-                    </div>
-                </div>
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <div class="alert alert-dismissible alert-light">
-                            <button type="button" class="btn-close" data-bs-dismiss="alert"></button>
-                            <strong>Heads up!</strong> This <a href="#" class="alert-link">alert needs your
-                                attention</a>,
-                            but it's not super important.
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div>
-                <h2>Badges</h2>
-                <div class="bs-component mb-4">
-                    <span class="badge bg-primary">Primary</span>
-                    <span class="badge bg-secondary">Secondary</span>
-                    <span class="badge bg-success">Success</span>
-                    <span class="badge bg-danger">Danger</span>
-                    <span class="badge bg-warning">Warning</span>
-                    <span class="badge bg-info">Info</span>
-                    <span class="badge bg-light">Light</span>
-                    <span class="badge bg-dark">Dark</span>
-                </div>
-                <div class="bs-component">
-                    <span class="badge rounded-pill bg-primary">Primary</span>
-                    <span class="badge rounded-pill bg-secondary">Secondary</span>
-                    <span class="badge rounded-pill bg-success">Success</span>
-                    <span class="badge rounded-pill bg-danger">Danger</span>
-                    <span class="badge rounded-pill bg-warning">Warning</span>
-                    <span class="badge rounded-pill bg-info">Info</span>
-                    <span class="badge rounded-pill bg-light">Light</span>
-                    <span class="badge rounded-pill bg-dark">Dark</span>
-                </div>
-            </div>
-        </div>
-
-        <!-- Progress
-        ================================================== -->
-        <div class="bs-docs-section">
-            <div class="row">
-                <div class="col-lg-12">
-                    <div class="page-header">
-                        <h1 id="progress">Progress</h1>
-                    </div>
-
-                    <h3 id="progress-basic">Basic</h3>
-                    <div class="bs-component">
-                        <div class="progress">
-                            <div class="progress-bar" role="progressbar" style="width: 25%;" aria-valuenow="25"
-                                 aria-valuemin="0" aria-valuemax="100"></div>
-                        </div>
-                    </div>
-
-                    <h3 id="progress-alternatives">Contextual alternatives</h3>
-                    <div class="bs-component">
-                        <div class="progress">
-                            <div class="progress-bar bg-success" role="progressbar" style="width: 25%;"
-                                 aria-valuenow="25"
-                                 aria-valuemin="0" aria-valuemax="100"></div>
-                        </div>
-                        <div class="progress">
-                            <div class="progress-bar bg-info" role="progressbar" style="width: 50%;" aria-valuenow="50"
-                                 aria-valuemin="0" aria-valuemax="100"></div>
-                        </div>
-                        <div class="progress">
-                            <div class="progress-bar bg-warning" role="progressbar" style="width: 75%;"
-                                 aria-valuenow="75"
-                                 aria-valuemin="0" aria-valuemax="100"></div>
-                        </div>
-                        <div class="progress">
-                            <div class="progress-bar bg-danger" role="progressbar" style="width: 100%;"
-                                 aria-valuenow="100"
-                                 aria-valuemin="0" aria-valuemax="100"></div>
-                        </div>
-                    </div>
-
-                    <h3 id="progress-multiple">Multiple bars</h3>
-                    <div class="bs-component">
-                        <div class="progress">
-                            <div class="progress-bar" role="progressbar" style="width: 15%;" aria-valuenow="15"
-                                 aria-valuemin="0" aria-valuemax="100"></div>
-                            <div class="progress-bar bg-success" role="progressbar" style="width: 30%;"
-                                 aria-valuenow="30"
-                                 aria-valuemin="0" aria-valuemax="100"></div>
-                            <div class="progress-bar bg-info" role="progressbar" style="width: 20%;" aria-valuenow="20"
-                                 aria-valuemin="0" aria-valuemax="100"></div>
-                        </div>
-                    </div>
-
-                    <h3 id="progress-striped">Striped</h3>
-                    <div class="bs-component">
-                        <div class="progress">
-                            <div class="progress-bar progress-bar-striped" role="progressbar" style="width: 10%;"
-                                 aria-valuenow="10" aria-valuemin="0" aria-valuemax="100"></div>
-                        </div>
-                        <div class="progress">
-                            <div class="progress-bar progress-bar-striped bg-success" role="progressbar"
-                                 style="width: 25%;"
-                                 aria-valuenow="25" aria-valuemin="0" aria-valuemax="100"></div>
-                        </div>
-                        <div class="progress">
-                            <div class="progress-bar progress-bar-striped bg-info" role="progressbar"
-                                 style="width: 50%;"
-                                 aria-valuenow="50" aria-valuemin="0" aria-valuemax="100"></div>
-                        </div>
-                        <div class="progress">
-                            <div class="progress-bar progress-bar-striped bg-warning" role="progressbar"
-                                 style="width: 75%;"
-                                 aria-valuenow="75" aria-valuemin="0" aria-valuemax="100"></div>
-                        </div>
-                        <div class="progress">
-                            <div class="progress-bar progress-bar-striped bg-danger" role="progressbar"
-                                 style="width: 100%;"
-                                 aria-valuenow="100" aria-valuemin="0" aria-valuemax="100"></div>
-                        </div>
-                    </div>
-
-                    <h3 id="progress-animated">Animated</h3>
-                    <div class="bs-component">
-                        <div class="progress">
-                            <div class="progress-bar progress-bar-striped progress-bar-animated" role="progressbar"
-                                 aria-valuenow="75" aria-valuemin="0" aria-valuemax="100" style="width: 75%;"></div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <!-- Containers
-        ================================================== -->
-        <div class="bs-docs-section">
-            <div class="row">
-                <div class="col-lg-12">
-                    <div class="page-header">
-                        <h1 id="containers">Containers</h1>
-                    </div>
-                </div>
-            </div>
-
-            <div class="row">
-                <div class="col-lg-12">
-                    <h2>List groups</h2>
-                </div>
-            </div>
-
-            <div class="row">
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <ul class="list-group">
-                            <li class="list-group-item d-flex justify-content-between align-items-center">
-                                Cras justo odio
-                                <span class="badge bg-primary rounded-pill">14</span>
-                            </li>
-                            <li class="list-group-item d-flex justify-content-between align-items-center">
-                                Dapibus ac facilisis in
-                                <span class="badge bg-primary rounded-pill">2</span>
-                            </li>
-                            <li class="list-group-item d-flex justify-content-between align-items-center">
-                                Morbi leo risus
-                                <span class="badge bg-primary rounded-pill">1</span>
-                            </li>
-                        </ul>
-                    </div>
-                </div>
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <div class="list-group">
-                            <a href="#" class="list-group-item list-group-item-action active">Cras justo odio</a>
-                            <a href="#" class="list-group-item list-group-item-action">Dapibus ac facilisis in</a>
-                            <a href="#" class="list-group-item list-group-item-action disabled">Morbi leo risus</a>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <div class="list-group">
-                            <a href="#"
-                               class="list-group-item list-group-item-action flex-column align-items-start active">
-                                <div class="d-flex w-100 justify-content-between">
-                                    <h5 class="mb-1">List group item heading</h5>
-                                    <small>3 days ago</small>
-                                </div>
-                                <p class="mb-1">Donec id elit non mi porta gravida at eget metus. Maecenas sed diam eget
-                                    risus varius blandit.</p>
-                                <small>Donec id elit non mi porta.</small>
-                            </a>
-                            <a href="#" class="list-group-item list-group-item-action flex-column align-items-start">
-                                <div class="d-flex w-100 justify-content-between">
-                                    <h5 class="mb-1">List group item heading</h5>
-                                    <small class="text-muted">3 days ago</small>
-                                </div>
-                                <p class="mb-1">Donec id elit non mi porta gravida at eget metus. Maecenas sed diam eget
-                                    risus varius blandit.</p>
-                                <small class="text-muted">Donec id elit non mi porta.</small>
-                            </a>
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-            <div class="row">
-                <div class="col-lg-12">
-                    <h2>Cards</h2>
-                </div>
-            </div>
-
-            <div class="row">
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <div class="card text-white bg-primary mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Primary card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card text-white bg-secondary mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Secondary card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card text-white bg-success mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Success card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card text-white bg-danger mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Danger card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card bg-warning mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Warning card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card text-white bg-info mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Info card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card bg-light mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Light card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card text-white bg-dark mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Dark card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <div class="card border-primary mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Primary card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card border-secondary mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Secondary card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card border-success mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Success card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card border-danger mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Danger card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card border-warning mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Warning card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card border-info mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Info card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card border-light mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Light card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                        <div class="card border-dark mb-1"
-                             style="max-width: 20rem;">
-                            <div class="card-header">Header</div>
-                            <div class="card-body">
-                                <h4 class="card-title">Dark card title</h4>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <div class="card mb-1">
-                            <h3 class="card-header">Card header</h3>
-                            <div class="card-body">
-                                <h5 class="card-title">Special title treatment</h5>
-                                <h6 class="card-subtitle text-muted">Support card subtitle</h6>
-                            </div>
-                            <svg xmlns="http://www.w3.org/2000/svg" class="d-block user-select-none" width="100%"
-                                 height="200" aria-label="Placeholder: Image cap" focusable="false" role="img"
-                                 preserveAspectRatio="xMidYMid slice" viewBox="0 0 318 180"
-                                 style="font-size:1.125rem;text-anchor:middle">
-                                <rect width="100%" height="100%" fill="#868e96"></rect>
-                                <text x="50%" y="50%" fill="#dee2e6" dy=".3em">Image cap</text>
-                            </svg>
-                            <div class="card-body">
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                            </div>
-                            <ul class="list-group list-group-flush">
-                                <li class="list-group-item">Cras justo odio</li>
-                                <li class="list-group-item">Dapibus ac facilisis in</li>
-                                <li class="list-group-item">Vestibulum at eros</li>
-                            </ul>
-                            <div class="card-body">
-                                <a href="#" class="card-link">Card link</a>
-                                <a href="#" class="card-link">Another link</a>
-                            </div>
-                            <div class="card-footer text-muted">
-                                2 days ago
-                            </div>
-                        </div>
-                        <div class="card">
-                            <div class="card-body">
-                                <h4 class="card-title">Card title</h4>
-                                <h6 class="card-subtitle mb-2 text-muted">Card subtitle</h6>
-                                <p class="card-text">Some quick example text to build on the card title and make up the
-                                    bulk
-                                    of the card's content.</p>
-                                <a href="#" class="card-link">Card link</a>
-                                <a href="#" class="card-link">Another link</a>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-            <div class="row mt-5">
-                <div class="col-lg-12">
-                    <h2>Accordions</h2>
-                </div>
-            </div>
-
-            <div class="row">
-                <div class="col-lg-4">
-                    <div class="bs-component">
-                        <div class="accordion"
-                             id="accordionExample">
-                            <div class="accordion-item">
-                                <h2 class="accordion-header"
-                                    id="headingOne">
-                                    <button class="accordion-button"
-                                            type="button"
-                                            data-bs-toggle="collapse"
-                                            data-bs-target="#collapseOne"
-                                            aria-expanded="true"
-                                            aria-controls="collapseOne">
-                                        Accordion Item #1
-                                    </button>
-                                </h2>
-                                <div id="collapseOne"
-                                     class="accordion-collapse collapse show"
-                                     aria-labelledby="headingOne"
-                                     data-bs-parent="#accordionExample">
-                                    <div class="accordion-body">
-                                        <strong>This is the first item's accordion body.</strong> It
-                                        is
-                                        shown
-                                        by
-                                        default,
-                                        until
-                                        the
-                                        collapse
-                                        plugin
-                                        adds
-                                        the
-                                        appropriate
-                                        classes
-                                        that
-                                        we
-                                        use
-                                        to
-                                        style
-                                        each
-                                        element.
-                                        These
-                                        classes
-                                        control
-                                        the
-                                        overall
-                                        appearance,
-                                        as
-                                        well
-                                        as
-                                        the
-                                        showing
-                                        and
-                                        hiding
-                                        via
-                                        CSS
-                                        transitions.
-                                        You
-                                        can
-                                        modify
-                                        any
-                                        of
-                                        this
-                                        with
-                                        custom
-                                        CSS
-                                        or
-                                        overriding
-                                        our
-                                        default
-                                        variables.
-                                        It's
-                                        also
-                                        worth
-                                        noting
-                                        that
-                                        just
-                                        about
-                                        any
-                                        HTML
-                                        can
-                                        go
-                                        within
-                                        the
-                                        <code>.accordion-body</code>, though the transition does limit
-                                        overflow.
-                                    </div>
-                                </div>
-                            </div>
-                            <div class="accordion-item">
-                                <h2 class="accordion-header"
-                                    id="headingTwo">
-                                    <button class="accordion-button collapsed"
-                                            type="button"
-                                            data-bs-toggle="collapse"
-                                            data-bs-target="#collapseTwo"
-                                            aria-expanded="false"
-                                            aria-controls="collapseTwo">
-                                        Accordion Item #2
-                                    </button>
-                                </h2>
-                                <div id="collapseTwo"
-                                     class="accordion-collapse collapse"
-                                     aria-labelledby="headingTwo"
-                                     data-bs-parent="#accordionExample">
-                                    <div class="accordion-body">
-                                        <strong>This is the second item's accordion body.</strong> It
-                                        is
-                                        hidden
-                                        by
-                                        default,
-                                        until
-                                        the
-                                        collapse
-                                        plugin
-                                        adds
-                                        the
-                                        appropriate
-                                        classes
-                                        that
-                                        we
-                                        use
-                                        to
-                                        style
-                                        each
-                                        element.
-                                        These
-                                        classes
-                                        control
-                                        the
-                                        overall
-                                        appearance,
-                                        as
-                                        well
-                                        as
-                                        the
-                                        showing
-                                        and
-                                        hiding
-                                        via
-                                        CSS
-                                        transitions.
-                                        You
-                                        can
-                                        modify
-                                        any
-                                        of
-                                        this
-                                        with
-                                        custom
-                                        CSS
-                                        or
-                                        overriding
-                                        our
-                                        default
-                                        variables.
-                                        It's
-                                        also
-                                        worth
-                                        noting
-                                        that
-                                        just
-                                        about
-                                        any
-                                        HTML
-                                        can
-                                        go
-                                        within
-                                        the
-                                        <code>.accordion-body</code>, though the transition does limit
-                                        overflow.
-                                    </div>
-                                </div>
-                            </div>
-                            <div class="accordion-item">
-                                <h2 class="accordion-header"
-                                    id="headingThree">
-                                    <button class="accordion-button collapsed"
-                                            type="button"
-                                            data-bs-toggle="collapse"
-                                            data-bs-target="#collapseThree"
-                                            aria-expanded="false"
-                                            aria-controls="collapseThree">
-                                        Accordion Item #3
-                                    </button>
-                                </h2>
-                                <div id="collapseThree"
-                                     class="accordion-collapse collapse"
-                                     aria-labelledby="headingThree"
-                                     data-bs-parent="#accordionExample">
-                                    <div class="accordion-body">
-                                        <strong>This is the third item's accordion body.</strong> It
-                                        is
-                                        hidden
-                                        by
-                                        default,
-                                        until
-                                        the
-                                        collapse
-                                        plugin
-                                        adds
-                                        the
-                                        appropriate
-                                        classes
-                                        that
-                                        we
-                                        use
-                                        to
-                                        style
-                                        each
-                                        element.
-                                        These
-                                        classes
-                                        control
-                                        the
-                                        overall
-                                        appearance,
-                                        as
-                                        well
-                                        as
-                                        the
-                                        showing
-                                        and
-                                        hiding
-                                        via
-                                        CSS
-                                        transitions.
-                                        You
-                                        can
-                                        modify
-                                        any
-                                        of
-                                        this
-                                        with
-                                        custom
-                                        CSS
-                                        or
-                                        overriding
-                                        our
-                                        default
-                                        variables.
-                                        It's
-                                        also
-                                        worth
-                                        noting
-                                        that
-                                        just
-                                        about
-                                        any
-                                        HTML
-                                        can
-                                        go
-                                        within
-                                        the
-                                        <code>.accordion-body</code>, though the transition does limit
-                                        overflow.
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-
-        </div>
-
-        <!-- Dialogs
-        ================================================== -->
-        <div class="bs-docs-section">
-            <div class="row">
-                <div class="col-lg-12">
-                    <div class="page-header">
-                        <h1 id="dialogs">Dialogs</h1>
-                    </div>
-                </div>
-            </div>
-            <div class="row">
-                <div class="col-lg-6">
-                    <h2>Modals</h2>
-                    <div class="bs-component">
-                        <div class="modal">
-                            <div class="modal-dialog" role="document">
-                                <div class="modal-content">
-                                    <div class="modal-header">
-                                        <h5 class="modal-title">Modal title</h5>
-                                        <button type="button" class="btn-close" data-bs-dismiss="modal"
-                                                aria-label="Close">
-                                            <span aria-hidden="true"></span>
-                                        </button>
-                                    </div>
-                                    <div class="modal-body">
-                                        <p>Modal body text goes here.</p>
-                                    </div>
-                                    <div class="modal-footer">
-                                        <button type="button" class="btn btn-primary">Save changes</button>
-                                        <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close
-                                        </button>
-                                    </div>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div class="col-lg-6">
-                    <h2>Popovers</h2>
-                    <div class="bs-component mb-5">
-                        <button type="button" class="btn btn-secondary" title="Popover Title" data-bs-container="body"
-                                data-bs-toggle="popover" data-bs-placement="left"
-                                data-bs-content="Vivamus sagittis lacus vel augue laoreet rutrum faucibus.">Left
-                        </button>
-
-                        <button type="button" class="btn btn-secondary" title="Popover Title" data-bs-container="body"
-                                data-bs-toggle="popover" data-bs-placement="top"
-                                data-bs-content="Vivamus sagittis lacus vel augue laoreet rutrum faucibus.">Top
-                        </button>
-
-                        <button type="button" class="btn btn-secondary" title="Popover Title" data-bs-container="body"
-                                data-bs-toggle="popover" data-bs-placement="bottom"
-                                data-bs-content="Vivamus sagittis lacus vel augue laoreet rutrum faucibus.">Bottom
-                        </button>
-
-                        <button type="button" class="btn btn-secondary" title="Popover Title" data-bs-container="body"
-                                data-bs-toggle="popover" data-bs-placement="right"
-                                data-bs-content="Vivamus sagittis lacus vel augue laoreet rutrum faucibus.">Right
-                        </button>
-                    </div>
-                    <h2>Tooltips</h2>
-                    <div class="bs-component mb-5">
-                        <button type="button" class="btn btn-secondary" data-bs-toggle="tooltip"
-                                data-bs-placement="left"
-                                title="Tooltip on left">Left
-                        </button>
-
-                        <button type="button" class="btn btn-secondary" data-bs-toggle="tooltip" data-bs-placement="top"
-                                title="Tooltip on top">Top
-                        </button>
-
-                        <button type="button" class="btn btn-secondary" data-bs-toggle="tooltip"
-                                data-bs-placement="bottom"
-                                title="Tooltip on bottom">Bottom
-                        </button>
-
-                        <button type="button" class="btn btn-secondary" data-bs-toggle="tooltip"
-                                data-bs-placement="right"
-                                title="Tooltip on right">Right
-                        </button>
-                    </div>
-                    <h2>Toasts</h2>
-                    <div class="bs-component">
-                        <div class="toast show" role="alert" aria-live="assertive" aria-atomic="true">
-                            <div class="toast-header">
-                                <strong class="me-auto">Bootstrap</strong>
-                                <small>11 mins ago</small>
-                                <button type="button" class="btn-close ms-2 mb-1" data-bs-dismiss="toast"
-                                        aria-label="Close">
-                                    <span aria-hidden="true"></span>
-                                </button>
-                            </div>
-                            <div class="toast-body">
-                                Hello, world! This is a toast message.
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <!-- Ícones do projeto
-        ================================================== -->
-        <div class="bs-docs-section">
-            <h2>
-                Ícones
-            </h2>
-            <div class="row g-1 mb-2 bg-light">
-                <div class="col-auto">
-                    <div class="bs-component">
-                        <x-frontend::icon class="ic-lg text-primary" name="fmd-logo"/>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <!-- Imagens para trocar
-        ================================================== -->
-        <div class="bs-docs-section">
-
-            <div class="row">
-                <div class="col-lg-12">
-                    <div class="page-header">
-                        <h1 id="dialogs">
-                            <a href="https://realfavicongenerator.net/"
-                               aria-label="Images"
-                               target="_blank">Images
-                            </a>
-                        </h1>
-                        <div class="d-flex align-items-center mb-1">
-                            <div class="bg-danger mr-0h"
-                                 style="width: 20px; height: 20px;"></div>
-                            Missing <kbd class="ml-0h">alt=""</kbd>
-                        </div>
-                        <div class="d-flex align-items-center mb-2">
-                            <div class="bg-warning mr-0h"
-                                 style="width: 20px; height: 20px"></div>
-                            Missing <kbd class="ml-0h">loading=""</kbd>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="row">
-                <div class="col-lg-12">
-                    <div class="page-header">
-                        <h1 id="links-warnings">
-                            <a href="https://realfavicongenerator.net/"
-                               target="_blank">Links
-                            </a>
-                        </h1>
-                        <div class="d-flex align-items-center mb-1">
-                            <div class="bg-danger mr-0h"
-                                 style="width: 20px; height: 20px;"></div>
-                            Missing <kbd class="ml-0h">aria-label=""</kbd>
-                        </div>
-                        <div class="d-flex align-items-center mb-2">
-                            <div class="bg-warning mr-0h"
-                                 style="width: 20px; height: 20px"></div>
-                            Missing <kbd class="ml-0h">title=""</kbd>
-                        </div>
-                    </div>
-                </div>
-            </div>
-            <div class="row">
-                <div class="col-6">
-                    <div class="p-3 mb-2 bg-light">
-                        <figure class="figure mb-0">
-                            <img src="/images/logo.png"
-                                 class="figure-img img-fluid"
-                                 alt="200x70 - logo.png (1,8 KB)"
-                                 loading="lazy"
-                                 width="200"
-                                 height="70">
-                            <figcaption class="figure-caption">200x70 - logo.png (1,8 KB)
-                            </figcaption>
-                        </figure>
-                    </div>
-                </div>
-                <div class="col-6">
-                    <div class="p-3 mb-2 bg-light">
-                        <figure class="figure mb-0">
-                            <img src="/images/logo-email.png"
-                                 class="figure-img img-fluid"
-                                 alt="200x70 - logo-email.png (1,8 KB)"
-                                 loading="lazy"
-                                 width="200"
-                                 height="70">
-                            <figcaption class="figure-caption">200x70 - logo-email.png (1,8
-                                KB)
-                            </figcaption>
-                        </figure>
-                    </div>
-                </div>
-                <div class="col-6">
-                    <div class="p-3 mb-2 bg-light">
-                        <figure class="figure mb-0">
-                            <img src="/images/favicon.ico"
-                                 class="figure-img img-fluid"
-                                 alt="48x48 - favicon.ico (15,09 KB)"
-                                 loading="lazy"
-                                 width="48"
-                                 height="48">
-                            <figcaption class="figure-caption">48x48 - favicon.ico (15,09
-                                KB)
-                            </figcaption>
-                        </figure>
-                    </div>
-                </div>
-                <div class="col-6">
-                    <div class="p-3 mb-2 bg-light">
-                        <figure class="figure mb-0">
-                            <img src="/images/logo-amp.png"
-                                 class="figure-img img-fluid"
-                                 alt="60x60 - logo-amp.png (1 KB)"
-                                 loading="lazy"
-                                 width="60"
-                                 height="60">
-                            <figcaption class="figure-caption">60x60 - logo-amp.png (1 KB)
-                            </figcaption>
-                        </figure>
-                    </div>
-                </div>
-                <div class="col-12">
-                    <div class="p-3 mb-2 bg-light">
-                        <figure class="figure mb-0">
-                            <img src="/images/mstile-150x150.png"
-                                 class="figure-img img-fluid"
-                                 alt="150x150 - mstile-150x150.png (1,7 KB)"
-                                 loading="lazy"
-                                 loading="lazy"
-                                 width="150"
-                                 height="150">
-                            <figcaption class="figure-caption">150x150 - mstile-150x150.png
-                                (1,7
-                                KB)
-                            </figcaption>
-                        </figure>
-                    </div>
-                </div>
-                <div class="col-6">
-                    <div class="p-3 mb-2 bg-light">
-                        <figure class="figure mb-0">
-                            <img src="/images/safari-pinned-tab.svg"
-                                 class="figure-img img-fluid"
-                                 alt="192x192 - safari-pinned-tab.svg (5 KB)"
-                                 loading="lazy"
-                                 width="192"
-                                 height="192">
-                            <figcaption class="figure-caption">192x192 -
-                                safari-pinned-tab.svg
-                                (5 KB)
-                            </figcaption>
-                        </figure>
-                    </div>
-                </div>
-                <div class="col-6">
-                    <div class="p-3 mb-2 bg-light">
-                        <figure class="figure mb-0">
-                            <img src="/images/apple-touch-icon.png"
-                                 class="figure-img img-fluid"
-                                 alt="192x192 - apple-touch-icon.png (3,51 KB)"
-                                 loading="lazy"
-                                 width="192"
-                                 height="192">
-                            <figcaption class="figure-caption">192x192 -
-                                apple-touch-icon.png
-                                (3,51 KB)
-                            </figcaption>
-                        </figure>
-                    </div>
-                </div>
-                <div class="col-12">
-                    <div class="p-3 mb-2 bg-light">
-                        <figure class="figure mb-0">
-                            <img src="/images/icon-192x192.png"
-                                 class="figure-img img-fluid"
-                                 alt="192x192 - icon-192x192.png (5,28 KB)"
-                                 loading="lazy"
-                                 width="192"
-                                 height="192">
-                            <figcaption class="figure-caption">192x192 - icon-192x192.png
-                                (5,28
-                                KB)
-                            </figcaption>
-                        </figure>
-                    </div>
-                </div>
-                <div class="col-12">
-                    <div class="p-3 mb-2 bg-light">
-                        <figure class="figure mb-0">
-                            <img src="/images/icon-512x512.png"
-                                 class="figure-img img-fluid"
-                                 alt="512x512 - icon-512x512.png (13,55 KB)"
-                                 loading="lazy"
-                                 width="512"
-                                 height="512">
-                            <figcaption class="figure-caption">
-                                512x512 - icon-512x512.png (13,55 KB)
-                            </figcaption>
-                        </figure>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <!--Inifinte Scroll Loading Icon-->
-        <h2>Ícone do Infinite Scroll</h2>
-        <div class="ic-infinite-scroll-loading"></div>
-
-        <div id="source-modal" class="modal fade" tabindex="-1">
-            <div class="modal-dialog modal-lg">
-                <div class="modal-content">
-                    <div class="modal-header">
-                        <h4 class="modal-title">Source Code</h4>
-                        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-hidden="true"></button>
-                    </div>
-                    <div class="modal-body">
-                        <pre class="language-html"><code></code></pre>
-                    </div>
-                </div>
-            </div>
-        </div>
-
-        <footer id="footer">
-            <div class="row">
-                <div class="col-lg-12">
-
-                    <ul class="list-unstyled">
-                        <li class="float-end">
-                            <a href="#top">Back to top</a>
-                        </li>
-                        <li>
-                            <a href="https://blog.bootswatch.com/">Blog</a>
-                        </li>
-                        <li>
-                            <a href="https://blog.bootswatch.com/rss/">RSS</a>
-                        </li>
-                        <li>
-                            <a href="https://twitter.com/bootswatch">Twitter</a>
-                        </li>
-                        <li>
-                            <a href="https://github.com/thomaspark/bootswatch">GitHub</a>
-                        </li>
-                        <li>
-                            <a href="../help/#api">API</a>
-                        </li>
-                        <li>
-                            <a href="../help/#donate">Donate</a>
-                        </li>
-                    </ul>
-                    <p>Made by
-                        <a href="https://thomaspark.co/">Thomas Park</a>
-                        .
-                    </p>
-                    <p>Code released under the
-                        <a href="https://github.com/thomaspark/bootswatch/blob/master/LICENSE">
-                            MIT
-                            License
-                        </a>
-                        .
-                    </p>
-                    <p>Based on
-                        <a href="https://getbootstrap.com/"
-                           rel="nofollow">Bootstrap
-                        </a>
-                        . Icons from
-                        <a href="https://fontawesome.com/"
-                           rel="nofollow">Font Awesome
-                        </a>
-                        . Web fonts from
-                        <a
-                                href="https://fonts.google.com/"
-                                rel="nofollow">Google
-                        </a>
-                        .
-                    </p>
-
-                </div>
-            </div>
-        </footer>
-    </main>
-
-@endsection
 @push('scripts')
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/prism/1.23.0/prism.min.js" data-manual></script>
 
-    <script>
-        (function () {
-            'use strict';
+@vite('resources/js/pages/theme.js')
 
-            $(window).scroll(function () {
-                var top = $(document).scrollTop();
-                if (top > 50) {
-                    $('#home > .navbar').removeClass('navbar-transparent');
-                } else {
-                    $('#home > .navbar').addClass('navbar-transparent');
-                }
-            });
-
-            $('a[href="#"]').click(function (event) {
-                event.preventDefault();
-            });
-
-            $('.bs-component').each(function () {
-                var $component = $(this);
-                var $button = $('<button class="source-button btn btn-primary btn-xs" role="button" tabindex="0">&lt; &gt;</button>');
-                $component.append($button);
-
-                if ($component.find('[data-bs-toggle="tooltip"]').length > 0) {
-                    $component.attr('data-html', $component.html().trim());
-                }
-            });
-
-            var popoverTriggerList = [].slice.call(document.querySelectorAll(
-                '[data-bs-toggle="popover"]'));
-            var popoverList = popoverTriggerList.map(function (popoverTriggerEl) {
-                return new bootstrap.Popover(popoverTriggerEl);
-            });
-
-            var tooltipTriggerList = [].slice.call(document.querySelectorAll(
-                '[data-bs-toggle="tooltip"]'));
-            var tooltipList = tooltipTriggerList.map(function (tooltipTriggerEl) {
-                return new bootstrap.Tooltip(tooltipTriggerEl);
-            });
-
-            var sourceModalElem = document.getElementById('source-modal');
-            if (sourceModalElem) {
-                var sourceModal = new bootstrap.Modal(document.getElementById('source-modal'));
-            }
-
-            $('body').on('click', '.source-button', function (event) {
-                event.preventDefault();
-
-                var component = $(this).parent();
-                var html = component.attr('data-html') ? component.attr('data-html') : component.html();
-
-                html = cleanSource(html);
-                html = Prism.highlight(html, Prism.languages.html, 'html');
-                $('#source-modal code').html(html);
-                sourceModal.show();
-            });
-
-            function cleanSource(html) {
-                html = html.replace(/×/g, '&times;')
-                    .replace(/«/g, '&laquo;')
-                    .replace(/»/g, '&raquo;')
-                    .replace(/←/g, '&larr;')
-                    .replace(/→/g, '&rarr;');
-
-                var lines = html.split(/\n/);
-
-                lines.shift();
-                lines.splice(-1, 1);
-
-                var indentSize = lines[0].length - lines[0].trim().length;
-                var re = new RegExp(' {' + indentSize + '}');
-
-                lines = lines.map(function (line) {
-                    if (line.match(re)) {
-                        line = line.slice(Math.max(0, indentSize));
-                    }
-
-                    return line;
-                });
-
-                lines = lines.join('\n');
-
-                return lines;
-            }
-        })();
-    </script>
 @endpush
