+++ resources/js/pages/theme.js
@@ -1,0 +1,7 @@
+import { setupShareAPI } from '../frontend.js';
+
+(function () {
+
+  setupShareAPI();
+
+})();
