+++ resources/scss/components/_icons.scss
@@ -1,21 +1,55 @@
-$icon-sm: 2rem !default;
-$icon-md: 3rem !default;
-$icon-lg: 4rem !default;
+$icon-xxxsm: 1rem !default; // 16
+$icon-xxsm: 1.25rem !default; //20
+$icon-xsm: 1.5rem !default; //24
+$icon-sm: 2rem !default; //32
+$icon-md: 3rem !default; //48
+$icon-lg: 3.25rem !default; //52
+$icon-xl: 4rem !default; //64
+
+@mixin icons-size-setup {
+
+  .ic {
+    max-width: 100%;
+    width: $icon-md;
+    height: $icon-md;
+
+    &-xxxsm {
+      width: $icon-xxxsm;
+      height: $icon-xxxsm;
+    }
+
+    &-xxsm {
+      width: $icon-xxsm;
+      height: $icon-xxsm;
+    }
+
+    &-xsm {
+      width: $icon-xsm;
+      height: $icon-xsm;
+    }
+
+    &-sm {
+      width: $icon-sm;
+      height: $icon-sm;
+    }
+
+    &-lg {
+      width: $icon-lg;
+      height: $icon-lg;
+    }
+
+    &-xl {
+      width: $icon-xl;
+      height: $icon-xl;
+    }
+  }
+}
+
+@import 'icon-link';
+
+@include icons-size-setup;
 
 .ic {
-  max-width: 100%;
-  width: $icon-md;
-  height: $icon-md;
-
-  &-sm {
-    width: $icon-sm;
-    height: $icon-sm;
-  }
-
-  &-lg {
-    width: $icon-lg;
-    height: $icon-lg;
-  }
 
   &-rotate-n90 {
     transform: rotate(-90deg);
@@ -35,15 +69,5 @@ $icon-lg: 4rem !default;
 
   &-inverse-y {
     transform: scaleY(-1);
-  }
-
-  &-infinite-scroll-loading {
-    $infinite-scroll-loading-icon-size: 64px !default;
-    $infinite-scroll-loading-icon-indicator-color: $primary !default;
-    $infinite-scroll-loading-icon-indicator: url("data:image/svg+xml,<svg xmlns='http://www.w3.org/2000/svg' version='1.0' width='#{$infinite-scroll-loading-icon-size}' height='#{$infinite-scroll-loading-icon-size}' viewBox='0 0 128 128' fill='#{$infinite-scroll-loading-icon-indicator-color}'><g><circle cx='16' cy='64' r='16'/><circle cx='16' cy='64' r='14.344' transform='rotate(45 64 64)'/><circle cx='16' cy='64' r='12.531' transform='rotate(90 64 64)'/><circle cx='16' cy='64' r='10.75' transform='rotate(135 64 64)'/><circle cx='16' cy='64' r='10.063' transform='rotate(180 64 64)'/><circle cx='16' cy='64' r='8.063' transform='rotate(225 64 64)'/><circle cx='16' cy='64' r='6.438' transform='rotate(270 64 64)'/><circle cx='16' cy='64' r='5.375' transform='rotate(315 64 64)'/><animateTransform attributeName='transform' type='rotate' values='0 64 64;315 64 64;270 64 64;225 64 64;180 64 64;135 64 64;90 64 64;45 64 64' calcMode='discrete' dur='1040ms' repeatCount='indefinite'/></g></svg>") !default;
-    $infinite-scroll-loading-icon-background: escape-svg($infinite-scroll-loading-icon-indicator) no-repeat center !default;
-
-    min-height: $infinite-scroll-loading-icon-size;
-    background: $infinite-scroll-loading-icon-background;
   }
 }
