+++ resources/scss/frontend.scss
@@ -81,6 +81,7 @@
 @import "components/navbar";
 @import "components/videos";
 @import "components/badges";
+@import "components/dropdown";
 @import "components/list";
 @import "styles";
