+++ resources/scss/frontend/tokens/_design-tokens.scss
@@ -49,4 +49,4 @@ $ds-gradient: linear-gradient(128.65deg, tint-color($ds-primary, $ds-shade-level
 
 //body
 $ds-body-color: $ds-dark;
-$ds-body-bg: $ds-primary;
+$ds-body-bg: $ds-white;
